% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_gm.R
\name{agents_players}
\alias{agents_players}
\title{Players Agents}
\usage{
agents_players(agents = "Jeff Schwartz", all_agents = F,
  return_message = FALSE)
}
\arguments{
\item{agents}{vector of agents}

\item{all_agents}{if \code{TRUE} scrapes all agents}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a  \code{data_frame}
}
\description{
Acquires NBA players agents
}
\examples{
agents_players(agents = c("Jeff Schwartz"))

}
\seealso{
Other agents: \code{\link{players_agents}}
}
\concept{agents}
