% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{all_star_games}
\alias{all_star_games}
\title{NBA All Star Games}
\usage{
all_star_games(include_aba = T, return_message = T)
}
\arguments{
\item{include_aba}{if `TRUE` includes ABA scores}

\item{return_message}{if `TRUE` returns a message}
}
\value{
a `tibble`
}
\description{
Returns scores from All-Star Games
}
\examples{
library(dplyr)
library(nbastatR)
df_asg <-
all_star_games()
df_asg \%>\% glimpse()
df_asg \%>\% count(namePlayerMVP, sort = T)
}
