% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{assign_bref_data}
\alias{assign_bref_data}
\title{Assign nested BREF data to environment}
\usage{
assign_bref_data(
  data,
  type = "Players",
  widen_data = TRUE,
  include_all_nba = F,
  join_data = TRUE,
  nest_data = FALSE,
  assign_to_environment = TRUE
)
}
\arguments{
\item{data}{a \code{tibble} of tables}

\item{type}{type of BREF data are `teams` and `players`}

\item{widen_data}{if \code{TRUE} widens data}

\item{include_all_nba}{if `TRUE` include all NBA teams}

\item{join_data}{if \code{TRUE} joins tables}

\item{nest_data}{if \code{TRUE} nests data}

\item{assign_to_environment}{if \code{TRUE} assigns data to your environment}
}
\value{
a `tibble`
}
\description{
Assign nested BREF data to environment
}
