% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ish.R
\name{assign_nba_players}
\alias{assign_nba_players}
\title{Assign NBA player dictionary to environment}
\usage{
assign_nba_players()
}
\value{
a `data_frame` to `df_dict_nba_players` in your environment
}
\description{
Assign NBA player dictionary to environment
}
\examples{
assign_nba_players()
}
\seealso{
Other dictionary: \code{\link{assign_nba_teams}},
  \code{\link{dictionary_bref_coaches}},
  \code{\link{dictionary_bref_players}},
  \code{\link{dictionary_bref_teams}}

Other NBA players: \code{\link{dictionary_bref_players}}
}
\concept{NBA players}
\concept{dictionary}
