% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{beyond_the_numbers}
\alias{beyond_the_numbers}
\title{NBA.com Beyond The Numbers Articles}
\usage{
beyond_the_numbers(count_articles = 50)
}
\arguments{
\item{count_articles}{numeric vector of counts}
}
\value{
a \code{data_frame}
}
\description{
Returns articles from beyond the numbers
}
\examples{
\dontrun{
beyond_the_numbers(count_articles = 10)
}
}
\seealso{
Other news: \code{\link{players_rotowire}},
  \code{\link{teams_rotowire}}
}
\concept{news}
