% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_scores.R
\name{box_scores}
\alias{box_scores}
\title{NBA box scores}
\usage{
box_scores(
  game_ids = NULL,
  league = "NBA",
  box_score_types = c("Traditional", "Advanced", "Scoring", "Misc", "Usage",
    "Four Factors", "hustle", "tracking"),
  result_types = c("player", "team"),
  join_data = TRUE,
  assign_to_environment = TRUE,
  return_message = TRUE
)
}
\arguments{
\item{game_ids}{vector of game ids}

\item{box_score_types}{vector of box score types options include \itemize{
\item traditional
\item advanced
\item scoring
\item misc
\item usage
\item four factors
\item tracking
\item defense
\item matchups
\item hustle
}}

\item{result_types}{vector of result types options include \itemize{
\item team - Team statistics
\item player - player
}}

\item{join_data}{if \code{TRUE} joins the underlying table data}

\item{assign_to_environment}{if \code{TRUE} assigns a data frame for each table to the environment
starting with data}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{tibble}
}
\description{
Acquires specified box score type by game ID
}
\examples{
box_scores(game_ids = c(21700002, 21700003), box_score_types = c("Traditional", "Advanced", "Scoring", "Misc", "Usage", "Four Factors", "Tracking"), result_types = c("player", "team"), join_data = TRUE, assign_to_environment = TRUE, return_message = TRUE)
}
\seealso{
Other game: 
\code{\link{fanduel_summary}()},
\code{\link{game_logs}()},
\code{\link{win_probability}()}
}
\concept{game}
