% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{bref_awards_votes}
\alias{bref_awards_votes}
\title{Basketball Reference award votes}
\usage{
bref_awards_votes(
  seasons = 2017,
  assign_to_environment = TRUE,
  return_message = TRUE
)
}
\arguments{
\item{seasons}{vector of seasons starting in 1964}

\item{assign_to_environment}{if \code{TRUE} assigns a data frame for each
nested table into your global environment}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{tibble()}
}
\description{
Acquires seasons awards votes
}
\examples{
bref_awards_votes(2017)
}
\seealso{
Other BREF: 
\code{\link{bref_awards}()},
\code{\link{bref_bios}()},
\code{\link{bref_players_stats}()},
\code{\link{bref_teams_stats}()},
\code{\link{dictionary_bref_coaches}()},
\code{\link{dictionary_bref_players}()},
\code{\link{dictionary_bref_teams}()}

Other awards: 
\code{\link{bref_awards}()},
\code{\link{players_awards}()}
}
\concept{BREF}
\concept{awards}
