% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref_bio.R
\name{bref_bios}
\alias{bref_bios}
\title{Basketball Reference players bios}
\usage{
bref_bios(players = NULL, player_ids = NULL,
  assign_to_environment = TRUE, return_message = T)
}
\arguments{
\item{players}{vector of player names}

\item{player_ids}{vector of basketball reference player ids}

\item{assign_to_environment}{if \code{TRUE} assigns each table to environment}

\item{return_message}{if \code{TRUE} returns}
}
\value{
a \code{tibble}
}
\description{
Includes player bios, salaries, contracts and transactions
}
\examples{
\dontrun{
bref_bios( players = c("Jarrett Allen", "Mitch Richmond", "Michael Adams"),
player_ids = NULL,
assign_to_environment = TRUE)
}
}
\seealso{
Other BREF: \code{\link{bref_awards_votes}},
  \code{\link{bref_awards}},
  \code{\link{bref_players_stats}},
  \code{\link{bref_teams_stats}},
  \code{\link{dictionary_bref_coaches}},
  \code{\link{dictionary_bref_players}},
  \code{\link{dictionary_bref_teams}}

Other player: \code{\link{draft_combines}},
  \code{\link{players_awards}},
  \code{\link{players_careers}},
  \code{\link{players_tables}}

Other salaries: \code{\link{hoops_hype_salary_summary}},
  \code{\link{hoopshype_salaries}},
  \code{\link{nba_insider_salaries}}
}
\concept{BREF}
\concept{player}
\concept{salaries}
