% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{bref_injuries}
\alias{bref_injuries}
\title{Active injuries}
\usage{
bref_injuries()
}
\value{
a \code{data_frame()}
}
\description{
Returns information of
active NBA player injuries from
Basketball-Reference
}
\examples{
bref_injuries()
}
