% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{bref_players_stats}
\alias{bref_players_stats}
\title{Basketball Reference Player Season Tables}
\usage{
bref_players_stats(seasons = NULL, tables = c("advanced", "totals"),
  include_all_nba = F, only_totals = TRUE, nest_data = FALSE,
  assign_to_environment = TRUE, widen_data = TRUE, join_data = TRUE,
  return_message = TRUE)
}
\arguments{
\item{seasons}{vector of years 1951 to current season}

\item{tables}{player table \itemize{
\item \code{totals}: Totals
\item \code{per_game}: Per game
\item \code{advanced}: Advanced
\item \code{per_minute}: Per 36 minutes
\item \code{per_poss}: Per Possession
}}

\item{include_all_nba}{if \code{TRUE} include all_nba teams}

\item{only_totals}{if \code{TRUE} returns only a player's total statistics}

\item{nest_data}{if \code{TRUE} returns a nested data frame}

\item{assign_to_environment}{if `TRUE` assigns to environment}

\item{widen_data}{if `TRUE` widens data}

\item{join_data}{if `TRUE` joins `tibbles`}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{tibble}
}
\description{
Basketball Reference Player Season Tables
}
\examples{
bref_players_stats(seasons = 2017:2018, tables = c("advanced", "totals"))
}
\seealso{
Other BREF: \code{\link{bref_awards_votes}},
  \code{\link{bref_awards}}, \code{\link{bref_bios}},
  \code{\link{bref_teams_stats}},
  \code{\link{dictionary_bref_coaches}},
  \code{\link{dictionary_bref_players}},
  \code{\link{dictionary_bref_teams}}
}
\concept{BREF}
\concept{player statistics}
