% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{coaching_staffs}
\alias{coaching_staffs}
\title{NBA active coaching staffs}
\usage{
coaching_staffs()
}
\value{
a \code{tibbles}
}
\description{
Gets active coaching staff information for all
NBA teams
}
\examples{
coaching_staffs()
}
\seealso{
Other Coaching: \code{\link{dictionary_bref_coaches}}
}
\concept{Coaching}
\concept{Current data}
\concept{Roster information}
