% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{current_schedule}
\alias{current_schedule}
\title{NBA current season schedule}
\usage{
current_schedule()
}
\value{
a `tibble`
}
\description{
Get NBA schedule for most current season
}
\examples{
current_schedule()
}
