% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{current_standings}
\alias{current_standings}
\title{Current standings}
\usage{
current_standings(return_message = TRUE)
}
\arguments{
\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a `tibble`
}
\description{
Acquires current season's standings
}
\examples{
current_standings()
}
\seealso{
Other standings: 
\code{\link{standings}()}
}
\concept{standings}
