% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{dictionary_bref_awards}
\alias{dictionary_bref_awards}
\title{Basketball Reference Awards}
\usage{
dictionary_bref_awards()
}
\value{
a \code{data_frame}
}
\description{
Basketball Reference Awards
}
\examples{
dictionary_bref_awards()
}
