% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{dictionary_bref_teams}
\alias{dictionary_bref_teams}
\title{Basketball Reference team dictionary}
\usage{
dictionary_bref_teams()
}
\value{
a `tibble`
}
\description{
Basketball Reference team dictionary
}
\examples{
dictionary_bref_teams()
}
\seealso{
Other dictionary: 
\code{\link{assign_nba_players}()},
\code{\link{assign_nba_teams}()},
\code{\link{dictionary_bref_coaches}()},
\code{\link{dictionary_bref_players}()}

Other BREF: 
\code{\link{bref_awards}()},
\code{\link{bref_awards_votes}()},
\code{\link{bref_bios}()},
\code{\link{bref_players_stats}()},
\code{\link{bref_teams_stats}()},
\code{\link{dictionary_bref_coaches}()},
\code{\link{dictionary_bref_players}()}
}
\concept{BREF}
\concept{dictionary}
