% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ish.R
\name{dictionary_nba_names}
\alias{dictionary_nba_names}
\title{Dictionary of NBA Headers and nbastatR names}
\usage{
dictionary_nba_names()
}
\value{
a \code{data_frame}
}
\description{
Dictionary of NBA Headers and nbastatR names
}
\examples{
dictionary_nba_names()
}
\concept{Dictionary}
\concept{nbastatR tool}
