% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{dictionary_player_photos}
\alias{dictionary_player_photos}
\title{Cached player photo dictionary}
\usage{
dictionary_player_photos()
}
\value{
\code{tibble}
}
\description{
Cached player photo dictionary
}
