% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_combine.R
\name{draft_combines}
\alias{draft_combines}
\title{NBA draft combine data}
\usage{
draft_combines(years = NULL, return_message = T, nest_data = F)
}
\arguments{
\item{years}{vector of draft years}

\item{return_message}{if \code{TRUE} return message}

\item{nest_data}{if \code{TRUE} returns nested tibble}
}
\value{
a \code{tibble()}
}
\description{
Acquires NBA draft combine data
}
\examples{
draft_combines(years = c(2001:2018),
nest_data = T)
}
\seealso{
Other draft: 
\code{\link{drafts}()},
\code{\link{nbadraftnet_mock_drafts}()}

Other player: 
\code{\link{bref_bios}()},
\code{\link{players_awards}()},
\code{\link{players_careers}()},
\code{\link{players_tables}()}
}
\concept{draft}
\concept{player}
