% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drafts.R
\name{drafts}
\alias{drafts}
\title{NBA drafts}
\usage{
drafts(draft_years = 1947:2018, nest_data = F, return_message = T)
}
\arguments{
\item{draft_years}{\code{vector} of draft years}

\item{nest_data}{if \code{TRUE} returns nested data frame}

\item{return_message}{if \code{TRUE} returns message}
}
\value{
a `data_frame`
}
\description{
NBA draft data for specified seasons
}
\examples{
library(dplyr)
df_drafts <-
drafts(draft_years = 1983:2018, nest_data = FALSE, return_message = TRUE)

## Where do top 5 picks since 1983 come from?

df_drafts \%>\%
filter(numberPickOverall <= 5) \%>\%
count(nameOrganizationFrom, sort = T)
}
\seealso{
Other draft: \code{\link{draft_combines}},
  \code{\link{nbadraftnet_mock_drafts}}

Other teams: \code{\link{franchise_leaders}},
  \code{\link{seasons_rosters}},
  \code{\link{teams_players_stats}},
  \code{\link{teams_rankings}}
}
\concept{draft}
\concept{teams}
