% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_by_play.R
\name{fanduel_summary}
\alias{fanduel_summary}
\title{Games fanduel summary}
\usage{
fanduel_summary(game_ids = c(21700002, 21700003), nest_data = FALSE,
  return_message = TRUE)
}
\arguments{
\item{game_ids}{vector of game ids}

\item{nest_data}{if \code{TRUE} nests data}

\item{return_message}{if \code{T} returns message}
}
\value{
a \code{data_frame}
}
\description{
Returns fanduel summary for specified
game ids
}
\examples{
fanduel_summary(game_ids = c(21700002, 21700003), nest_data = F, return_message = T)
}
\seealso{
Other season: \code{\link{win_probability}}

Other game: \code{\link{box_scores}},
  \code{\link{game_logs}}, \code{\link{win_probability}}
}
\concept{game}
\concept{season}
