% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_names}
\alias{fix_names}
\title{Fix data frame's names}
\usage{
fix_names(data, slugs = c("FD", "DK"))
}
\arguments{
\item{data}{\code{data_frame}}

\item{slugs}{vector of slugs}
}
\description{
Fix data frame's names
}
