% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_leaders.R
\name{franchise_leaders}
\alias{franchise_leaders}
\title{Franchise leaders}
\usage{
franchise_leaders(teams = NULL, all_teams = FALSE,
  remove_inactive_teams = F, modes = "Totals",
  season_types = c("Regular Season"), return_message = TRUE,
  nest_data = FALSE)
}
\arguments{
\item{teams}{vector of team names}

\item{all_teams}{if \code{TRUE} returns all teams}

\item{remove_inactive_teams}{if \code{TRUE} removes inactive teams}

\item{modes}{mode of search \itemize{
\item Totals
\item Per Game
}}

\item{season_types}{type of season \itemize{
\item Regular Season
\item Playoffs
\item Pre Season
}'}
}
\value{
a \code{tibble}
}
\description{
Gets franchise leader information for
by specified input for specific teams
}
\examples{
franchise_leaders(teams = "Brooklyn Nets", modes = c("Totals"))
}
\seealso{
Other teams: \code{\link{drafts}},
  \code{\link{seasons_rosters}},
  \code{\link{teams_players_stats}},
  \code{\link{teams_rankings}}

Other leaders: \code{\link{metrics_leaders}}
}
\concept{leaders}
\concept{teams}
