% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nba_players_ids.R
\docType{data}
\name{function_packages}
\alias{function_packages}
\title{Title}
\format{An object of class \code{character} of length 7.}
\usage{
function_packages
}
\arguments{
\item{active_only}{}
}
\description{
Title
}
\examples{

}
\keyword{datasets}

