% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logs.R
\name{game_logs}
\alias{game_logs}
\title{NBA game logs
NBA game logs for specified parameters}
\usage{
game_logs(seasons = 2019, league = "NBA", result_types = "player",
  season_types = "Regular Season", nest_data = F,
  assign_to_environment = TRUE, return_message = TRUE, ...)
}
\arguments{
\item{seasons}{vector of seasons where season is year ending}

\item{league}{league name defaults to WNBA}

\item{result_types}{vector of result types \itemize{
\item player
\item team
}}

\item{season_types}{vector of Season types \itemize{
\item Regular Season
\item Playoffs
\item Pre Season
\item All Star
}}

\item{nest_data}{if \code{TRUE} nests data}

\item{assign_to_environment}{assigns individual table to environment}

\item{return_message}{}
}
\value{
a `data_frame`
}
\description{
NBA game logs
NBA game logs for specified parameters
}
\examples{
game_logs(seasons = 2019, result_types = c("team", "player"))
}
\seealso{
Other game: \code{\link{box_scores}},
  \code{\link{fanduel_summary}},
  \code{\link{win_probability}}
}
\concept{game}
