% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gather_data}
\alias{gather_data}
\title{Gather a data frame}
\usage{
gather_data(
  data,
  variable_name = "item",
  numeric_ids = c("^id"),
  use_logical_keys = TRUE,
  use_factor_keys = TRUE,
  unite_columns = NULL,
  separate_columns = NULL,
  use_date_keys = FALSE,
  remove_na = TRUE
)
}
\arguments{
\item{data}{a \code{tibble}}

\item{variable_name}{variable column name}

\item{numeric_ids}{vector of numeric ids}

\item{use_logical_keys}{if \code{TRUE} uses logicals as keys}

\item{use_factor_keys}{if \code{TRUE} uses factors as a key}

\item{unite_columns}{if not \code{NULL} \code{list} \itemize{
\item new_column : new column name
\item column_1 : first column to unite
\item column_2 : second column to unite
\item sep : separator
}}

\item{separate_columns}{if not \code{NULL} \code{list} \itemize{
\item column : column to separate
\item new_column_1 : new_column 1
\item new_column_2 : new_column 2
\item sep : separator
}}

\item{use_date_keys}{if \code{TRUE} uses dates as a key}

\item{remove_na}{removes NA columns}
}
\value{
a \code{tibble}
}
\description{
Gather a data frame
}
