% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_som.R
\name{generate_som}
\alias{generate_som}
\title{Generate SOM}
\usage{
generate_som(data, normalization_parameters = list(minute_minimum = 250,
  scale_to_minute = TRUE, id_column = "idPlayerSeason", data_columns =
  c("totalsFG3M", "totalsFG3A", "totalsFG2M", "totalsFG2A", "totalsFTM",
  "totalsFTA", "totalsORB", "totalsDRB", "totalsAST", "totalsSTL", "totalsBLK",
  "totalsTOV", "totalsPF", "totalsPTS", "pct3PRate", "pctFTRate", "pctORB",
  "pctDRB", "pctAST", "pctSTL", "pctBLK", "pctTOV")),
  topology_parameters = list(x_dim = NULL, y_dim = NULL, topology =
  "hexagonal", seed = 998, clusters = 10, rlength = 1000, include_neural_gas =
  TRUE, override_max_nodes = TRUE), filter_na = TRUE,
  assign_to_environment = TRUE, return_message = TRUE)
}
\arguments{
\item{data}{a \code{data_frame}}

\item{normalization_parameters}{a \code{list} containing \itemize{
\item minute minimum
\item scale_to_minute
\item id_column
\item data columns
}}

\item{topology_parameters}{a \code {list} containing \itemize{
\item x_dim
\item y_dim
\item seed
\item clusters
\item rlength
\item include_neural_gas
\item include_neighbors
\item override max_nodes
}}

\item{return_message}{}
}
\description{
Generate SOM
}
