% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_gm.R
\name{get_agents_player_data_df}
\alias{get_agents_player_data_df}
\title{Get agents players}
\usage{
get_agents_player_data_df(agents = "Jeff Schwartz", all_agents = FALSE,
  return_message = FALSE)
}
\arguments{
\item{agents}{vector of agents}

\item{all_agents}{if \code{TRUE} scrapes all agents}

\item{message}{}
}
\description{
Get agents players
}
