% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_gm.R
\name{get_agents_players}
\alias{get_agents_players}
\title{Get agents players}
\usage{
get_agents_players(agents = "Jeff Schwartz", all_agents = T,
  return_message = FALSE)
}
\arguments{
\item{agents}{vector of agents}

\item{all_agents}{if \code{TRUE} scrapes all agents}

\item{message}{}
}
\description{
Get agents players
}
