% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_insider_functions.R
\name{get_all_nba_teams_salaries}
\alias{get_all_nba_teams_salaries}
\title{Get All NBA Teams Salaries}
\usage{
get_all_nba_teams_salaries(assume_player_opt_out = T,
  assume_team_doesnt_exercise = T, spread_data = F, return_message = T)
}
\arguments{
\item{assume_team_doesnt_exercise}{}
}
\description{
Get All NBA Teams Salaries
}
