% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{get_all_star_game_scores}
\alias{get_all_star_game_scores}
\title{NBA All Star Games}
\usage{
get_all_star_game_scores(include_aba = T, return_message = T)
}
\arguments{
\item{include_aba}{if `TRUE` includes ABA scores}

\item{return_message}{if `TRUE` returns a message}
}
\value{
a `data_frame`
}
\description{
Returns scores from All-Star Games
}
\examples{
library(dplyr)
library(nbastatR)
df_asg <-
get_all_star_game_scores()
df_asg \%>\% glimpse()
df_asg \%>\% count(namePlayerMVP, sort = T)
}
