% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_rosters_history.R
\name{get_all_teams_seasons_coaches}
\alias{get_all_teams_seasons_coaches}
\title{Get All Team Seasons Coaches}
\usage{
get_all_teams_seasons_coaches(seasons = 2005:2006, return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get All Team Seasons Coaches
}

