% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{get_beyond_the_numbers_articles}
\alias{get_beyond_the_numbers_articles}
\title{NBA.com Beyond The Numbers Articles}
\usage{
get_beyond_the_numbers_articles(count_articles = 50)
}
\arguments{
\item{count_articles}{numeric vector of counts}
}
\value{
a \code{data_frame}
}
\description{
Returns articles from beyond the numbers
}
\examples{
\dontrun{
get_beyond_the_numbers_articles(count_aricles = 10)
}
}
\seealso{
Other news: \code{\link{get_players_roto_wire_news}},
  \code{\link{get_teams_roto_wire_news}}
}
\concept{news}
