% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{get_bref_active_injuries}
\alias{get_bref_active_injuries}
\title{Active injuries}
\usage{
get_bref_active_injuries()
}
\value{
a \code{data_frame()}
}
\description{
Returns information of
active NBA player injuries from
Basketball-Reference
}
\examples{
get_bref_active_injuries()
}
