% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{get_bref_all_nba_teams}
\alias{get_bref_all_nba_teams}
\title{Returns All NBA Teams}
\usage{
get_bref_all_nba_teams(return_message = TRUE)
}
\arguments{
\item{return_message}{if \code{TRUE} returns a message}
}
\description{
Returns All NBA Teams
}
\examples{
get_bref_all_nba_teams()
}
