% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{get_bref_player_dictionary}
\alias{get_bref_player_dictionary}
\title{Basketball Reference Player Dictionary}
\usage{
get_bref_player_dictionary()

get_bref_player_dictionary()
}
\value{
a `data_frame`
}
\description{
Basketball reference player dictionary
}
\examples{
get_bref_player_dictionary()
get_bref_player_dictionary()
}
