% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref_bio.R
\name{get_bref_players_bios}
\alias{get_bref_players_bios}
\title{Basketball Reference players bios}
\usage{
get_bref_players_bios(players = NULL, player_ids = NULL,
  assign_to_environment = TRUE, return_message = T)
}
\arguments{
\item{players}{vector of player names}

\item{player_ids}{vector of basketball reference player ids}

\item{assign_to_environment}{if \code{TRUE} assigns each table to environment}

\item{return_message}{if \code{TRUE} returns}
}
\value{
a \code{data_frame}
}
\description{
Includes player bios, salaries, contracts and transactions
}
\examples{
\dontrun{
get_bref_players_bios( players = c("Jarrett Allen", "Mitch Richmond", "Michael Adams"),
player_ids = NULL,
assign_to_environment = TRUE)
}
}
\seealso{
Other BREF: \code{\link{bref_awards_votes}},
  \code{\link{bref_awards}},
  \code{\link{bref_players_seasons}},
  \code{\link{bref_teams_seasons}},
  \code{\link{dictionary_bref_coaches}},
  \code{\link{dictionary_bref_players}},
  \code{\link{dictionary_bref_teams}}

Other player: \code{\link{get_players_awards}},
  \code{\link{get_players_career_stats}},
  \code{\link{get_players_tables_data}},
  \code{\link{get_years_draft_combines}}

Other salaries: \code{\link{get_all_nba_teams_salaries}},
  \code{\link{get_hoops_hype_teams_salary_summary}},
  \code{\link{get_nba_team_salaries}},
  \code{\link{get_teams_hoopshype_salaries}}
}
\concept{BREF}
\concept{player}
\concept{salaries}
