% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{get_bref_players_seasons}
\alias{get_bref_players_seasons}
\title{Basketball Reference Player Season Tables}
\usage{
get_bref_players_seasons(seasons = 2016:2018, tables = c("advanced",
  "totals"), include_all_nba = TRUE, only_totals = TRUE,
  nest_data = FALSE, assign_to_environment = TRUE, widen_data = TRUE,
  join_data = TRUE, return_message = TRUE)
}
\arguments{
\item{seasons}{vector of years 1951 to current season}

\item{tables}{player table \itemize{
\item \code{totals}: Totals
\item \code{per_game}: Per game
\item \code{advanced}: Advanced
\item \code{per_minute}: Per 36 minutes
\item \code{per_poss}: Per Possesion
}}

\item{include_all_nba}{if \code{TRUE} include all_nba teams}

\item{only_totals}{if \code{TRUE} returns only a player's total statistics}

\item{nest_data}{if \code{TRUE} returns a nested data frame}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{data_frame}
}
\description{
Basketball Reference Player Season Tables
}
\examples{
get_bref_players_seasons(seasons = 2018, tables = c("advanced", "totals"))
}
