% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{get_bref_seasons_award_votes}
\alias{get_bref_seasons_award_votes}
\title{Basketball Reference award votes}
\usage{
get_bref_seasons_award_votes(seasons = 2017, assign_to_environment = TRUE,
  return_message = TRUE)
}
\arguments{
\item{seasons}{vector of seasons starting in 1964}

\item{assign_to_environment}{if \code{TRUE} assigns a data frame for each
nested table into your global environment}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{data_frame()}
}
\description{
Acquires seasons awards votes
}
\examples{
get_bref_seasons_award_votes
}
\seealso{
Other BREF: \code{\link{get_bref_awards}},
  \code{\link{get_bref_coach_dictionary}},
  \code{\link{get_bref_player_dictionary}},
  \code{\link{get_bref_players_bios}},
  \code{\link{get_bref_players_seasons}},
  \code{\link{get_bref_team_dictionary}},
  \code{\link{get_bref_teams_seasons}}

Other awards: \code{\link{get_bref_all_nba_teams}},
  \code{\link{get_bref_awards}},
  \code{\link{get_players_awards}}
}
\concept{BREF}
\concept{awards}
