% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{get_bref_team_dictionary}
\alias{get_bref_team_dictionary}
\title{Basketball Reference team dictionary}
\usage{
get_bref_team_dictionary()
}
\value{
a `data_frame`
}
\description{
Basketball Reference team dictionary
}
\examples{
get_bref_team_dictionary()
}
\seealso{
Other dictionary: \code{\link{assign_nba_players}},
  \code{\link{assign_nba_teams}},
  \code{\link{get_bref_coach_dictionary}},
  \code{\link{get_bref_player_dictionary}}

Other BREF: \code{\link{get_bref_coach_dictionary}},
  \code{\link{get_bref_player_dictionary}},
  \code{\link{get_bref_players_bios}},
  \code{\link{get_bref_players_seasons}},
  \code{\link{get_bref_teams_seasons}}
}
\concept{BREF}
\concept{dictionary}
