% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{get_current_standings}
\alias{get_current_standings}
\title{Current standings}
\usage{
get_current_standings(return_message = TRUE)
}
\arguments{
\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a `data_frame`
}
\description{
Acquires current season's standings
}
\examples{
get_current_standings()
}
\seealso{
Other standings: \code{\link{get_seasons_standings}}
}
\concept{standings}
