% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{get_data_bref_players_seasons}
\alias{get_data_bref_players_seasons}
\title{Basketball Reference Player Season Tables}
\usage{
get_data_bref_players_seasons(tables = c("advanced", "totals"),
  years = 2017, only_totals = TRUE, nest_data = FALSE,
  return_message = TRUE)
}
\arguments{
\item{tables}{player table \itemize{
\item \code{totals}: Totals
\item \code{'per_game'}: Per game
\item \code{advanced}: Advanced
\item \code{per_minute}: Per 36 minutes
\item \code{per_poss}: Per Possesion
}}

\item{years}{vector of years 1951 to current season}

\item{only_totals}{if \code{TRUE} returns only a player's total statistics}

\item{nest_data}{if \code{TRUE} returns a nested data frame}

\item{return_message}{if \code{TRUE} returns a message}
}
\description{
Basketball Reference Player Season Tables
}
