% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{get_data_bref_teams_seasons}
\alias{get_data_bref_teams_seasons}
\title{BREF Team Season Data}
\usage{
get_data_bref_teams_seasons(years = 1950:2018, return_message = TRUE,
  assign_to_environment = TRUE, nest_data = FALSE, join_data = TRUE,
  widen_data = TRUE)
}
\arguments{
\item{years}{vector of years from 1950 to current year}

\item{return_message}{if \code{TRUE} returns message}

\item{assign_to_environment}{if \code{TRUE} assigns data}

\item{widen_data}{if \code{TRUE} returns a merged widend data}
}
\value{
a \code{data_frame} with a list of \code{data_frames}
}
\description{
Get team table data by seasons
}
