% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_day_nba_game_player_tracking_availability}
\alias{get_day_nba_game_player_tracking_availability}
\title{Get Day NBA Game Player Tracking Availibility}
\usage{
get_day_nba_game_player_tracking_availability(date, merge_team_data = T,
  return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get Day NBA Game Player Tracking Availibility
}
