% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_day_nba_game_scores}
\alias{get_day_nba_game_scores}
\title{Title}
\usage{
get_day_nba_game_scores(date, return_message = T, merge_data_team = T)
}
\arguments{
\item{merge_data_team}{}
}
\description{
Title
}

