% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_day_nba_game_team_leaders}
\alias{get_day_nba_game_team_leaders}
\title{Get Day NBA Game Team Leaders}
\usage{
get_day_nba_game_team_leaders(date = "03-10-2016", merge_team_data = T,
  return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get Day NBA Game Team Leaders
}

