% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_day_nba_games}
\alias{get_day_nba_games}
\title{Get Day NBA Game}
\usage{
get_day_nba_games(date, return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get Day NBA Game
}

