% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_day_nba_league_standings}
\alias{get_day_nba_league_standings}
\title{Get Day NBA League Standings}
\usage{
get_day_nba_league_standings(date, return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get Day NBA League Standings
}
