% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_day_nba_matchups}
\alias{get_day_nba_matchups}
\title{Get Day NBA Matchups}
\usage{
get_day_nba_matchups(date = "03-10-2016", is.today = F)
}
\arguments{
\item{date}{}
}
\description{
Get Day NBA Matchups
}
