% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_day_nba_teams_last_meeting}
\alias{get_day_nba_teams_last_meeting}
\title{Get Day NBA Tams Last Meeting}
\usage{
get_day_nba_teams_last_meeting(date, merge_team_data = T,
  return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get Day NBA Tams Last Meeting
}
