% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_days_nba_game_player_tracking_availability}
\alias{get_days_nba_game_player_tracking_availability}
\title{Get Days NBA Player Tracking Availibility Data}
\usage{
get_days_nba_game_player_tracking_availability(dates, merge = T,
  message = T)
}
\arguments{
\item{message}{}
}
\description{
Get Days NBA Player Tracking Availibility Data
}

