% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_days_nba_game_team_leaders}
\alias{get_days_nba_game_team_leaders}
\title{Get Days NBA Game Team Leaders}
\usage{
get_days_nba_game_team_leaders(dates, merge = T, message = T)
}
\arguments{
\item{message}{}
}
\description{
Get Days NBA Game Team Leaders
}
