% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_days_nba_games}
\alias{get_days_nba_games}
\title{Get Days NBA Games}
\usage{
get_days_nba_games(dates, message = T)
}
\arguments{
\item{message}{}
}
\description{
Get Days NBA Games
}
\examples{
get_days_nba_games(dates = c("03-10-2016", "02-10-2016"))
}
