% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_days_nba_league_standings}
\alias{get_days_nba_league_standings}
\title{Get Days NBA League Standings}
\usage{
get_days_nba_league_standings(dates, message = T)
}
\arguments{
\item{message}{}
}
\description{
Get Days NBA League Standings
}
