% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_days_nba_matchups}
\alias{get_days_nba_matchups}
\title{Get Days NBA Matchups}
\usage{
get_days_nba_matchups(dates, is.today = F, message = T)
}
\arguments{
\item{message}{}
}
\description{
Get Days NBA Matchups
}

