% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_days_nba_scores}
\alias{get_days_nba_scores}
\title{Get Days NBA Scores}
\usage{
get_days_nba_scores(dates, merge = T, message = T)
}
\arguments{
\item{message}{}
}
\description{
Get Days NBA Scores
}
