% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_days_nba_teams_last_meeting}
\alias{get_days_nba_teams_last_meeting}
\title{Get days NBA teams last meeting}
\usage{
get_days_nba_teams_last_meeting(dates, merge = T, message = T)
}
\arguments{
\item{message}{}
}
\description{
Get days NBA teams last meeting
}
