% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nba_player_career_stats.R
\name{get_fd_name_df}
\alias{get_fd_name_df}
\title{Title}
\usage{
get_fd_name_df()
}
\arguments{
\item{player}{, name of a player}

\item{id.player}{, id of a player}

\item{use_totals}{, use the totals for the seasoni if traded or each team}

\item{per_mode}{either Totals, Per 36, or Per Game}

\item{return_message}{}
}
\description{
Title
}
