% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logs.R
\name{get_game_logs}
\alias{get_game_logs}
\title{NBA game logs
NBA game logs for specified parameters}
\usage{
get_game_logs(seasons = NULL, result_types = NULL,
  season_types = "Regular Season", nest_data = F,
  assign_to_environment = TRUE, return_message = TRUE, ...)
}
\arguments{
\item{seasons}{vector of seasons where season is year ending}

\item{result_types}{vector of result types \itemize{
\item player
\item team
}}

\item{season_types}{vector of Season types \itemize{
\item Regular Season
\item Playoffs
\item Pre Season
\item All Star
}}

\item{nest_data}{if \code{TRUE} nests data}

\item{assign_to_environment}{assigns individual table to environment}

\item{...}{}
}
\value{
a `data_frame`
}
\description{
NBA game logs
NBA game logs for specified parameters
}
\examples{
get_game_logs(seasons = 2017:2018, result_types = c("team", "player"))
}
