% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_scores.R
\name{get_games_box_scores}
\alias{get_games_box_scores}
\title{Get boxscores by Game ID}
\usage{
get_games_box_scores(game_ids = c(21700002, 21700003),
  box_score_types = c("traditional", "advanced", "scoring", "misc", "usage",
  "four factors", "hustle", "tracking"), result_types = c("player", "team"),
  assign_to_environment = TRUE, return_message = TRUE)
}
\arguments{
\item{game_ids}{vector of game ids}

\item{box_score_types}{vector of box score types options include \itemize{
\item traditional
\item advanced
\item scoring
\item misc
\item usage
\item four factors
\item tracking
}}

\item{result_types}{vector of result types options include \itemize{
\item team - Team statistics
\item player - player
}}

\item{assign_to_environment}{if \code{TRUE} assigns a data frame for each table to the environment
starting with data}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{data_frame}
}
\description{
Get boxscores by Game ID
}
\examples{
get_games_play_by_play(game_ids = c(21700002, 21700003), box_score_types = c("traditional", "advanced", "scoring","misc", "usage", "four factors", "tracking"), result_types = c("player", "team"), nest_data = FALSE, assign_to_environment = TRUE, return_message = TRUE)
}
