% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_by_play.R
\name{get_games_fanduel_summary}
\alias{get_games_fanduel_summary}
\title{Get games' fanduel summary}
\usage{
get_games_fanduel_summary(game_ids = c(21700002, 21700003),
  nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{game_ids}{vector of game ids}

\item{nest_data}{if \code{TRUE} nests data}

\item{return_message}{if \code{T} returns message}
}
\value{
a \code{data_frame}
}
\description{
Get games' fanduel summary
}
\examples{
get_games_fanduel_summary(game_ids = c(21700002, 21700003), nest_data = F, return_message = T)
}
