% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoops_hype_functions.R
\name{get_hoops_hype_teams_salary_summary}
\alias{get_hoops_hype_teams_salary_summary}
\title{Get HoopsHype NBA teams salaries}
\usage{
get_hoops_hype_teams_salary_summary(spread_data = F, return_message = T)
}
\arguments{
\item{spread_data}{if \code{TRUE} spreads data}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{data_frame()}
}
\description{
Get HoopsHype NBA teams salaries
}
\examples{
get_hoops_hype_teams_salary_summary(spread_data = TRUE)
}
