% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{get_nba_all_star_game_scores}
\alias{get_nba_all_star_game_scores}
\title{NBA All Sta Game Scores}
\usage{
get_nba_all_star_game_scores(include_aba = T, return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
NBA All Sta Game Scores
}
