% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{get_nba_coaching_staffs}
\alias{get_nba_coaching_staffs}
\title{Get active NBA coaching staffs}
\usage{
get_nba_coaching_staffs()
}
\value{
a \code{data_frames}
}
\description{
Get active NBA coaching staffs
}
\examples{
get_nba_coaching_staffs()
}
