% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{get_nba_current_season_schedule}
\alias{get_nba_current_season_schedule}
\title{NBA current season schedule}
\usage{
get_nba_current_season_schedule()
}
\value{
a `data_frame`
}
\description{
Get NBA schedule for most current season
}
\examples{
get_nba_current_season_schedule()
}
