% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_standings.R
\name{get_nba_current_standings}
\alias{get_nba_current_standings}
\title{Get NBA Current Standings}
\usage{
get_nba_current_standings()
}
\description{
Get NBA Current Standings
}
\examples{
get_nba_current_standings()
}
