% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_score_functions.R
\name{get_nba_day_score_json_data}
\alias{get_nba_day_score_json_data}
\title{Get Days JSON Score data}
\usage{
get_nba_day_score_json_data(game_date = "04-10-2016")
}
\arguments{
\item{game_date}{}
}
\description{
Get Days JSON Score data
}

