% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drafts.R
\name{get_nba_drafts}
\alias{get_nba_drafts}
\title{Get NBA Drafts}
\usage{
get_nba_drafts(draft_years = 1947:2018, nest_data = F, return_message = T)
}
\arguments{
\item{draft_years}{\code{vector} of draft years}

\item{nest_data}{if \code{TRUE} returns nested data frame}

\item{return_message}{if \code{TRUE} returns message}
}
\description{
Get NBA Drafts
}
\examples{
get_nba_drafts(draft_years = 2000:2018, nest_data = FALSE, return_message = TRUE)
}
