% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nba_team_stats_tables.R
\name{get_nba_franchise_data}
\alias{get_nba_franchise_data}
\title{get_nba_franchise_data}
\usage{
get_nba_franchise_data(return_franchises = c("all", "active", "current"),
  return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
get_nba_franchise_data
}

