% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nba_injuries.R
\name{get_nba_player_injuries}
\alias{get_nba_player_injuries}
\title{NBA Injuries}
\usage{
get_nba_player_injuries(filter_returning_today = T)
}
\arguments{
\item{filter_returning_today}{if \code{TRUE} filter players returning today}
}
\description{
NBA Injuries
}
\examples{
get_nba_player_injuries()
}
