% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{get_nba_players}
\alias{get_nba_players}
\title{Get Players from NBA Stats API}
\usage{
get_nba_players()
}
\value{
a \code{data_frame}
}
\description{
Get Players from NBA Stats API
}
\examples{
get_nba_players()
}
