% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nba_player_career_stats.R
\name{get_nba_players_career_stats}
\alias{get_nba_players_career_stats}
\title{Get NBA Players Summary Statistics by Mode}
\usage{
get_nba_players_career_stats(player_ids = NULL, players = NULL,
  modes = c("PerGame", "Totals"), assign_to_environment = TRUE,
  add_mode_names = TRUE, return_message = TRUE)
}
\arguments{
\item{player_ids}{\code{vector} of NBA Player IDs}

\item{players}{\code{NULL} or \code{vector} of NBA players}

\item{modes}{\code{vector} of items that can include \itemize{
\item Totals
\item PerGame
\item Per36
}}

\item{assign_to_environment}{if \code{TRUE} assign to each table environment
with a name starting with data}

\item{add_mode_names}{if \code{TRUE} adds the type of mode to the table data}

\item{return_message}{if \code{TRUE} returns a messag}
}
\value{
a \code{data_frame}
}
\description{
Get NBA Players Summary Statistics by Mode
}
\examples{
get_nba_players_career_stats(players = c("Joe Harris", "Myles Turner"), modes = c("Totals", "PerGame"))
}
