% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_profile.R, R/split_functions.R
\name{get_nba_players_ids}
\alias{get_nba_players_ids}
\title{Get NBA Player IDs}
\usage{
get_nba_players_ids(league = "NBA", active_only = F)

get_nba_players_ids(league = "NBA", active_only = F)
}
\description{
Get NBA Player IDs

Get NBA Player IDs
}

