% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_profile.R
\name{get_nba_players_profiles}
\alias{get_nba_players_profiles}
\title{Get NBA players profiles}
\usage{
get_nba_players_profiles(player_ids = NULL, players = NULL, nest_data = F,
  return_message = TRUE)
}
\arguments{
\item{player_ids}{Vector of player IDs}

\item{players}{vector of player names}

\item{nest_data}{}
}
\description{
Get NBA players profiles
}
\examples{
get_nba_players_profiles(player_ids = c(203500, 1628384), players = c("Michael Jordan", "Caris LeVert", "Jarrett Allen"), nest_data = FALSE, return_message = TRUE)
}
