% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst.R
\name{get_nba_pst_transaction}
\alias{get_nba_pst_transaction}
\title{ProSports NBA Transcations}
\usage{
get_nba_pst_transaction(person = "Jarrett Jack", team = NULL,
  date_from = NULL, date_to = NULL, include_trades = T,
  include_injury_list_movement = T, include_g_league_movement = T,
  include_injury_missed_games = T, include_personal_missed_games = T,
  include_discipline = T, include_criminal_incidents = T,
  return_message = T)
}
\arguments{
\item{person}{if not \code{NULL} vector of person name}

\item{team}{if not \code{NULL} vector of team name}

\item{date_from}{if not \code{NULL} vector ofdate}

\item{date_to}{if not \code{NULL} vector of person name}

\item{include_trades}{if \code{TRUE} includs trades}

\item{include_injury_list_movement}{if \code{TRUE} includes injury list}

\item{include_g_league_movement}{if \code{TRUE} includes g-league movenment}

\item{include_injury_missed_games}{if \code{TRUE} includes injury missed games}

\item{include_personal_missed_games}{if \code{TRUE} inclue personal missed games}

\item{include_discipline}{if \code{TRUE} include disciple}

\item{include_criminal_incidents}{if \code{TRUE} include criminal incidents}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{data_frame}
}
\description{
Returns data matching specified
parameters
}
\examples{
get_nba_pst_transaction(person = NULL, team = "Nets")
}
