% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_game_logs.R
\name{get_nba_seasons_gamelogs}
\alias{get_nba_seasons_gamelogs}
\title{Get NBA game logs for specified parameters}
\usage{
get_nba_seasons_gamelogs(seasons = 2017:2018, result_types = "player",
  season_types = "Regular Season", nest_data = F, return_message = TRUE,
  ...)
}
\arguments{
\item{seasons}{vector of seasons where season is year ending}

\item{result_types}{vector of result types \itemize{
\item player
\item team
}}

\item{season_types}{vector of Season types \itemize{
\item Regular Season
\item Playoffs
\item Pre Season
\item All Star
}}

\item{nest_data}{if \code{TRUE} nests data}

\item{...}{}
}
\description{
Get NBA game logs for specified parameters
}
\examples{
get_nba_seasons_gamelogs(seasons = 2017:2018, result_types = "team")
}
