% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_rankings.R
\name{get_nba_team_current_rankings}
\alias{get_nba_team_current_rankings}
\title{Get NBA Team Rankings for most current period}
\usage{
get_nba_team_current_rankings()
}
\description{
Get NBA Team Rankings for most current period
}
