% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_insider_functions.R
\name{get_nba_team_salaries}
\alias{get_nba_team_salaries}
\title{NBA team salary}
\usage{
get_nba_team_salaries(team_name = "Brooklyn Nets", team_slug = NA,
  assume_player_opt_out = T, assume_team_doesnt_exercise = T,
  spread_data = FALSE, return_message = T)
}
\arguments{
\item{team_name}{NBA team name}

\item{team_slug}{NBA team slug}

\item{assume_player_opt_out}{if `TRUE` assumes NBA player opts out}

\item{assume_team_doesnt_exercise}{if `TRUE` assumes teams don't exercise team option}

\item{spread_data}{if `TRUE` returns wide data}

\item{return_message}{if `TRUE` returns a message}
}
\value{
a `data_frame`
}
\description{
Returns information about an NBA teams salary
}
\examples{
get_nba_team_salaries(team_name = "Brooklyn Nets")
}
\seealso{
Other salaries: \code{\link{get_all_nba_teams_salaries}},
  \code{\link{get_bref_players_bios}},
  \code{\link{get_hoops_hype_teams_salary_summary}},
  \code{\link{get_teams_hoopshype_salaries}}
}
