% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_rosters_history.R
\name{get_nba_team_season_coaches}
\alias{get_nba_team_season_coaches}
\title{Get NBA Team Season Coaches}
\usage{
get_nba_team_season_coaches(team, year_season_end = 2016,
  return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get NBA Team Season Coaches
}
