% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_rosters_history.R
\name{get_nba_team_season_roster}
\alias{get_nba_team_season_roster}
\title{Get NBA Team's Season Roster}
\usage{
get_nba_team_season_roster(team = "Denver Nuggets", season = 2015,
  return_message = T)
}
\arguments{
\item{team}{team name}

\item{season}{season vector}

\item{return_message}{}
}
\description{
Get NBA Team's Season Roster
}
\examples{
get_nba_team_season_roster(team = "Denver Nuggets", season = 1991)
}
