% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nba_team_stats_tables.R
\name{get_nba_team_stat_table}
\alias{get_nba_team_stat_table}
\title{get_nba_team_stat_table}
\usage{
get_nba_team_stat_table(year.season_start = 2015,
  season_type = c("Regular Season", "Pre Season", "Playoffs", "All Star"),
  measure_type = c("Base", "Advanced", "Misc", "Scoring", "Four Factors",
  "Opponent"), per_mode = c("Totals", "PerGame", "MinutesPer", "Per48",
  "Per40", "Per36", "PerMinute", "PerPossession", "PerPlay",
  "Per100Possessions", "Per100Plays"), is.pace_adjusted = F, period = 0,
  player_experience = c(NA, "Rookie", "Sophomore", "Veteran"),
  player_position = c(NA, "F", "C", "G"), is.rank = F,
  is.plus_minus = c(F, T), game_segment = c(NA, "First Half", "Second Half",
  "Overtime"), conference = c(NA, "East", "West"), division = c(NA,
  "Atlantic", "Central", "Northwest", "Pacific", "Southeast", "Southwest"),
  division_against = c(NA, "Atlantic", "Central", "Northwest", "Pacific",
  "Southeast", "Southwest"), conference_against = c(NA, "East", "West"),
  ahead_behind = c(NA, "Ahead", "Behind"), date_from = NA, date_to = NA,
  game_scope = c(NA, "Yesterday", "Last 10"), last_n_games = 0,
  location = c(NA, "Home", "Road"), month = 0, season_segment = c(NA,
  "Post All-Star", "Pre All-Star"), opponent = NA, team = NA,
  outcome = c(NA, "W", "L"), playoff_round = 0, shot_clock_range = c(NA,
  "24-22", "22-18 Very Early", "18-15 Early", "15-7 Average", "7-4 Late",
  "4-0 Very Late", "ShotClock Off"), starter_bench = c(NA, "Starters",
  "Bench"), return_metadata = F, include_measure_name = c(T, F),
  return_message = F, ...)
}
\arguments{
\item{...}{}
}
\description{
get_nba_team_stat_table
}

