% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ish.R
\name{get_nba_teams_ids}
\alias{get_nba_teams_ids}
\title{NBA team ids}
\usage{
get_nba_teams_ids(teams = NULL, team_ids = NULL, all_active_teams = F)
}
\arguments{
\item{teams}{vector of team names}

\item{team_ids}{vector of team ids}
}
\value{
a `data_frame`
}
\description{
Resolve NBA team ID
}
\examples{
get_nba_teams_ids(teams = c("Brooklyn Nets", "Denver Nuggets"))
}
\seealso{
Other ids: \code{\link{get_nba_players_ids}}
}
\concept{ids}
