% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{get_nba_teams_seasons}
\alias{get_nba_teams_seasons}
\title{NBA teams seasons}
\usage{
get_nba_teams_seasons()
}
\value{
a \code{data_frame()}
}
\description{
NBA teams seasons
}
\examples{
get_nba_teams_seasons()
}
