% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{get_nba_transactions}
\alias{get_nba_transactions}
\title{NBA transactions since 2012}
\usage{
get_nba_transactions(include_history = T)
}
\value{
a `data_frame`
}
\description{
Acquires all NBA transactions since 2012
}
\examples{
get_nba_transactions()
}
\concept{transactions}
