% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_profile.R
\name{get_player_profile}
\alias{get_player_profile}
\title{Get Player Profile}
\usage{
get_player_profile(player, id.player = NULL, include_headline_stat = T,
  return_message = T)
}
\description{
Get Player Profile
}
