% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_season_gamelog.R
\name{get_player_season_gamelog}
\alias{get_player_season_gamelog}
\title{Title}
\usage{
get_player_season_gamelog(player, id.player = NULL,
  season_type = "Regular Season", year.season_start,
  include_date_detail = T, include_player_metadata = F,
  return_message = T)
}
\arguments{
\item{year.season_start}{}
}
\description{
Title
}
\examples{
get_player_season_gamelog(player = "John Stockton", year.season_start = 1994, include_player_metadata = T)
}

