% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_season_shot_data.R
\name{get_player_season_shot_data}
\alias{get_player_season_shot_data}
\title{GET NBA PLAYER SHOT DETAIL DATA}
\usage{
get_player_season_shot_data(player = "Stephen Curry",
  year_season_end = 2016, return_df_only = F, use_shot_zone_side = F,
  season_type = "Regular Season", shots_type = c("Dunk", "Layup",
  "Jump Shot", "Fadeaway", "Bank", "Tip", "Hook"),
  shot_areas = c("Three Point", "Paint", "Mid-Range"), vs_conference = NA,
  vs_division = NA, quarter_range = 1:12, minute_range = 0:12,
  against_team = NA, game_location = NA, game_month = NA, outcome = NA,
  position = NA, season_segment = NA, exclude_backcourt = T,
  return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
GET NBA PLAYER SHOT DETAIL DATA
}
\examples{
get_player_season_shot_data(player = "K.J. Mcdaniels", year_season_end = 2015)
}
