% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_gm.R
\name{get_players_agents}
\alias{get_players_agents}
\title{Get NBA Players Agents}
\usage{
get_players_agents(nest_data = F, return_message = T)
}
\arguments{
\item{return_message}{if \code{TRUE} retruns player url}
}
\description{
Get NBA Players Agents
}
\examples{
get_players_agents()
}
