% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{get_players_roto_wire_news}
\alias{get_players_roto_wire_news}
\title{Players RotoWire news}
\usage{
get_players_roto_wire_news(players = NULL, player_ids = NULL,
  nest_data = F, results = 50, return_message = TRUE)
}
\arguments{
\item{players}{vector of player_names}

\item{player_ids}{vector of player names}

\item{nest_data}{if \code{TRUE} returns a nested data frame}

\item{results}{integer of results}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a `data_frame`
}
\description{
Returns rotowire news for specified
players.
}
\examples{
get_players_roto_wire_news(players = c( "Jarrett Allen", "DeMarre Carroll", "Allen Crabbe"))
}
\seealso{
Other news: \code{\link{get_beyond_the_numbers_articles}},
  \code{\link{get_teams_roto_wire_news}}
}
