% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_functions.R
\name{get_players_season_stat_splits}
\alias{get_players_season_stat_splits}
\title{Get Players Seasons Stat Splits}
\usage{
get_players_season_stat_splits(players = c("Brook Lopez", "Joe Johnson",
  "Kyle Lowry"), year.season_start = 2015, season_type = c("Regular Season",
  "Pre Season", "Playoffs", "All Star"), measure_type = c("Base", "Advanced",
  "Misc", "Scoring", "Four Factors", "Opponent"), per_mode = c("PerGame",
  "Totals", "MinutesPer", "Per48", "Per40", "Per36", "PerMinute",
  "PerPossession", "PerPlay", "Per100Possessions", "Per100Plays"),
  is.plus_minus = c(F, T), is.pace_adjusted = F, period = 0,
  is.rank = F, game_segment = c(NA, "First Half", "Second Half",
  "Overtime"), division_against = c(NA, "Atlantic", "Central", "Northwest",
  "Pacific", "Southeast", "Southwest"), conference_against = c(NA, "East",
  "West"), date_from = NA, date_to = NA, last_n_games = 0,
  location = c(NA, "Home", "Road"), month = 0, season_segment = c(NA,
  "Post All-Star", "Pre All-Star"), opponent = NA, outcome = c(NA, "W",
  "L"), playoff_round = 0, shot_clock_range = c(NA, "24-22",
  "22-18 Very Early", "18-15 Early", "15-7 Average", "7-4 Late",
  "4-0 Very Late", "ShotClock Off"), message = T)
}
\arguments{
\item{message}{}
}
\description{
Get Players Seasons Stat Splits
}

