% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{get_players_tables_data}
\alias{get_players_tables_data}
\title{Title}
\usage{
get_players_tables_data(players = NULL, player_ids = NULL,
  tables = c("year over year", "passes", "game splits"), measures = "Base",
  seasons = 2018, modes = c("PerGame", "Totals"),
  season_types = "Regular Season", playoff_rounds = NA, is_plus_minus = F,
  n_games = 20, is_rank = F, is_pace_adjusted = F, outcomes = NA,
  locations = NA, months = NA, season_segments = NA, date_from = NA,
  date_to = NA, opponent_ids = NA, vs_confs = NA, vs_divisions = NA,
  game_segments = NA, periods = NA, shot_clocks = NA, last_n_games = NA,
  assign_to_environment = T, return_message = TRUE)
}
\arguments{
\item{return_messages}{}
}
\description{
Title
}
