% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_leaders.R
\name{get_seasons_metrics_league_leaders}
\alias{get_seasons_metrics_league_leaders}
\title{Get League Leaders by Season by Type}
\usage{
get_seasons_metrics_league_leaders(seasons = 2017:2018, metric = "pts",
  season_types = "Regular Season", modes = c("PerGame", "Totals"),
  return_message = TRUE, nest_data = FALSE)
}
\arguments{
\item{seasons}{vector of seasons}

\item{metric}{name of metric to sort on \itemize{
\item pts
\item min
\item ast
\item treb
\item oreb
\item dreb
\item stl
\item tov
\item fg3
\item fga
\item fgm
\item fta
\item ftm
\item pf
\item eff

}}

\item{season_types}{type of season \itemize{
\item Regular Season
\item Playoffs
\item Pre Season
}}

\item{modes}{mode of search \itemize{
\item Totals
\item Per Game
\item Per48
}}

\item{return_message}{if \code{TRUE} returns a message}

\item{nest_data}{if \code{TRUE} returns a nested data frame}
}
\value{
a \code{data_frame}
}
\description{
Get League Leaders by Season by Type
}
\examples{
get_seasons_metrics_league_leaders(seasons = 2000:2005, metric = "pts", season_types = "Regular Season", modes = "PerGame", return_message = T)
}
