% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_seasons.R
\name{get_seasons_players}
\alias{get_seasons_players}
\title{Seasons players}
\usage{
get_seasons_players(seasons = 1960:2018, nest_data = F,
  return_message = T)
}
\arguments{
\item{seasons}{numeric vector of seasons}

\item{nest_data}{if \code{TRUE} returns a nested data_frame}

\item{return_message}{if \code{TRUE} return message}
}
\value{
a `data_frame`
}
\description{
Seasons players
}
\examples{
get_seasons_players(2010:2017, nest_data = T, return_message = T)
}
