% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{get_seasons_playoff_picture}
\alias{get_seasons_playoff_picture}
\title{NBA seasons playoff picture}
\usage{
get_seasons_playoff_picture(seasons = NULL, assign_to_environment = TRUE,
  include_numeric_records = F, nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{seasons}{vector of seasons}

\item{assign_to_environment}{if \code{TRUE} returns assigns a data frame for each table in environment}

\item{nest_data}{if \code{TRUE} returns nested data frame}

\item{return_message}{if \code{TRUE} returns a message}

\item{include_include_numeric_records}{if \code{TRUE} parses records}
}
\value{
a `data_frame`
}
\description{
Returns playoff picture for specified seasons
}
\examples{
get_seasons_playoff_picture(seasons = 2015:2018,
assign_to_environment = TRUE,
 include_numeric_records = T)
}
\concept{playoffs}
