% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logs.R
\name{get_seasons_rosters}
\alias{get_seasons_rosters}
\title{Get seasons rosters}
\usage{
get_seasons_rosters(seasons = 2000:2018, return_message = TRUE,
  nest_data = F)
}
\arguments{
\item{seasons}{vector of seasons}

\item{return_message}{if \code{TRUE} returns a message}

\item{nest_data}{if \code{TRUE} nests data}
}
\value{
a `date_frame`
}
\description{
Returns rosters for each team of a specified season
}
\examples{
get_seasons_rosters(2018)
}
