% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logs.R
\name{get_seasons_schedule}
\alias{get_seasons_schedule}
\title{Get NBA Seasons Schedule}
\usage{
get_seasons_schedule(seasons = 2018, season_types = "Regular Season",
  parse_boxscores = F, box_score_tables = c("Traditional"),
  nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{seasons}{vector of seasons where season is year ending}

\item{season_types}{type of season}

\item{nest_data}{}
}
\value{
a \code{data_frame()}
}
\description{
Get NBA Seasons Schedule
}
\examples{
get_seasons_schedule(seasons = c(2012, 2018))
}
