% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{get_seasons_standings}
\alias{get_seasons_standings}
\title{Get seasons standing data}
\usage{
get_seasons_standings(seasons = 1950:2018,
  season_types = c("Regular Season"), resolve_records = TRUE,
  nest_data = F, return_message = TRUE)
}
\arguments{
\item{seasons}{vector of seasons}

\item{season_types}{type of season options include \itemize{
\item Regular Season
\item Pre Season
}}

\item{resolve_records}{\code{TRUE} resolve records into wins, losses and win percentage}

\item{nest_data}{if \code{TRUE} returns a nested data frame}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a `data_frame`
}
\description{
NBA season standing data
}
\examples{
get_seasons_standings(seasons = 2015:2018, season_types = "Regular Season", resolve_records = T, nest_data = F, return_message = T)
}
