% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_rosters_history.R
\name{get_seasons_teams_coaches}
\alias{get_seasons_teams_coaches}
\title{Get Seaons Coaching Staffs}
\usage{
get_seasons_teams_coaches(seasons = 1990:2018, nest_data = F,
  return_message = T)
}
\arguments{
\item{seasons}{vector of seasons}

\item{nest_data}{if \code{TRUE} nests data}

\item{return_message}{if \code{TRUE} returns a message}
}
\description{
Get Seaons Coaching Staffs
}
\examples{
get_seasons_teams_coaches(1990:1992)
}
