% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_rosters_history.R
\name{get_team_coaches}
\alias{get_team_coaches}
\title{Get Team's Season Coaching Staff}
\usage{
get_team_coaches(team = "Denver Nuggets", season = 2015,
  return_message = T)
}
\arguments{
\item{team}{name of the team}

\item{season}{season}

\item{return_message}{if \code{TRUE} returns a message}
}
\description{
Get Team's Season Coaching Staff
}
\examples{
get_team_coaches("Brooklyn Nets", 2016)
}
