% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_rosters_history.R
\name{get_team_season_roster}
\alias{get_team_season_roster}
\title{Team roster}
\usage{
get_team_season_roster(team = "Denver Nuggets", season = 2015,
  return_message = T)
}
\arguments{
\item{team}{team name}

\item{season}{season vector}

\item{return_message}{}
}
\value{
a `data_frame`
}
\description{
Gets teams roster from a specified seasons
}
\examples{
get_team_season_roster(team = "Denver Nuggets", season = 1991)
}
\seealso{
Other rosters: \code{\link{get_seasons_rosters}},
  \code{\link{get_seasons_teams_rosters}}
}
\concept{rosters}
