% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_teams_season_shot_data.R
\name{get_team_season_shot_data}
\alias{get_team_season_shot_data}
\title{Get Teams Shot Data}
\usage{
get_team_season_shot_data(team = "Brooklyn Nets", year_roster = 2016,
  year_data = 2016, use_shot_zone_side = F,
  season_type = "Regular Season", positions = c("G", "F", "C"),
  shots_type = c("Dunk", "Layup", "Jump Shot", "Fadeaway", "Bank", "Tip",
  "Hook"), shot_areas = c("Three Point", "Paint", "Mid-Range"),
  vs_conference = NA, vs_division = NA, quarter_range = 1:12,
  minute_range = 0:12, against_team = NA, game_location = NA,
  game_month = NA, outcome = NA, season_segment = NA,
  exclude_backcourt = T, return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get Teams Shot Data
}

