% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_salaries.R
\name{get_team_yahoo_team_salary_data}
\alias{get_team_yahoo_team_salary_data}
\title{get_team_yahoo_player_salary_data}
\usage{
get_team_yahoo_team_salary_data(team_name = c("New York Knicks"),
  format_currency = T, nest_data = T, return_wide = F,
  return_message = T)
}
\arguments{
\item{return_wide}{}
}
\value{

}
\description{
get_team_yahoo_player_salary_data
}
