% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_tables.R
\name{get_teams_details}
\alias{get_teams_details}
\title{NBA teams details}
\usage{
get_teams_details(teams = NULL, team_ids = NULL, all_teams = F,
  assign_to_environment = TRUE, return_message = T)
}
\arguments{
\item{teams}{vector of team names}

\item{team_ids}{vector of team ids}

\item{all_teams}{if \code{TRUE} returns all team names}

\item{assign_to_environment}{if \code{TRUE} assigns each table to a data frame in
your environment starting with data}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a `data_frame`
}
\description{
Gets information about NBA franchises
}
\examples{
get_teams_details(all_teams = TRUE, assign_to_environment = TRUE)

}
