% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_leaders.R
\name{get_teams_franchise_leaders}
\alias{get_teams_franchise_leaders}
\title{Franchise leaders}
\usage{
get_teams_franchise_leaders(teams = NULL, all_teams = FALSE,
  remove_inactive_teams = F, modes = "Totals",
  season_types = c("Regular Season"), return_message = TRUE,
  nest_data = FALSE)
}
\arguments{
\item{teams}{vector of team names}

\item{all_teams}{if \code{TRUE} returns all teams}

\item{remove_inactive_teams}{if \code{TRUE} removes inactive teams}

\item{modes}{mode of search \itemize{
\item Totals
\item Per Game
}}

\item{season_types}{type of season \itemize{
\item Regular Season
\item Playoffs
\item Pre Season
}'}
}
\value{
a \code{data_frame}
}
\description{
Gets franchise leader information for
by specified input for specific teams
}
\examples{
get_teams_franchise_leaders(teams = "Brooklyn Nets", modes = c("Totals"))
}
\seealso{
Other teams: \code{\link{get_drafts}},
  \code{\link{get_seasons_rosters}},
  \code{\link{get_teams_players_seasons_summary_stats}},
  \code{\link{get_teams_seasons_rankings}}

Other leaders: \code{\link{get_seasons_metrics_league_leaders}}
}
\concept{leaders}
\concept{teams}
