% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{get_teams_roto_wire_news}
\alias{get_teams_roto_wire_news}
\title{Get teams roto wire news}
\usage{
get_teams_roto_wire_news(teams = NULL, nest_data = F, results = 50,
  return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Returns roto wire news for specified
teams.
}
