% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_functions.R
\name{get_teams_season_stat_splits}
\alias{get_teams_season_stat_splits}
\title{Get Teams Seasons Stat Splits}
\usage{
get_teams_season_stat_splits(teams = c("Atlanta Hawks", "Boston Celtics",
  "Brooklyn Nets", "Charlotte Hornets", "Chicago Bulls", "Cleveland Cavaliers",
  "Dallas Mavericks", "Denver Nuggets", "Detroit Pistons",
  "Golden State Warriors", "Houston Rockets", "Indiana Pacers",
  "Los Angeles Clippers", "Los Angeles Lakers", "Memphis Grizzlies",
  "Miami Heat", "Milwaukee Bucks", "Minnesota Timberwolves",
  "New Orleans Pelicans", "New York Knicks", "Oklahoma City Thunder",
  "Orlando Magic", "Philadelphia 76ers", "Phoenix Suns",
  "Portland Trail Blazers",      "Sacramento Kings", "San Antonio Spurs",
  "Toronto Raptors", "Utah Jazz", "Washington Wizards"),
  year.season_start = 2015, season_type = c("Regular Season", "Pre Season",
  "Playoffs", "All Star"), measure_type = c("Base", "Advanced", "Misc",
  "Scoring", "Four Factors", "Opponent"), per_mode = c("Totals", "PerGame",
  "MinutesPer", "Per48", "Per40", "Per36", "PerMinute", "PerPossession",
  "PerPlay", "Per100Possessions", "Per100Plays"), is.plus_minus = c(F, T),
  is.pace_adjusted = F, period = 0, is.rank = F, game_segment = c(NA,
  "First Half", "Second Half", "Overtime"), division_against = c(NA,
  "Atlantic", "Central", "Northwest", "Pacific", "Southeast", "Southwest"),
  conference_against = c(NA, "East", "West"), date_from = NA,
  date_to = NA, last_n_games = 0, location = c(NA, "Home", "Road"),
  month = 0, season_segment = c(NA, "Post All-Star", "Pre All-Star"),
  opponent = NA, outcome = c(NA, "W", "L"), playoff_round = 0,
  shot_clock_range = c(NA, "24-22", "22-18 Very Early", "18-15 Early",
  "15-7 Average", "7-4 Late", "4-0 Very Late", "ShotClock Off"), message = T)
}
\arguments{
\item{message}{}
}
\description{
Get Teams Seasons Stat Splits
}
