% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{get_teams_seasons_shots}
\alias{get_teams_seasons_shots}
\title{Get teams seasons shot charts}
\usage{
get_teams_seasons_shots(teams = NULL, team_ids = NULL,
  all_active_teams = F, season_types = "Regular Season",
  seasons = 2017:2018, measures = "FGA", periods = 0, months = 0,
  date_from = NA, date_to = NA, nest_data = F, return_message = T)
}
\arguments{
\item{return_message}{}
}
\value{
a \code{data_frame}
}
\description{
Get teams seasons shot charts
}
\examples{
get_teams_seasons_shots(teams = "Brooklyn Nets", seasons = 2018)
}
