% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{get_teams_tables_data}
\alias{get_teams_tables_data}
\title{Title}
\usage{
get_teams_tables_data(teams = NULL, team_ids = NULL, all_active_teams = F,
  tables = c("performance", "splits", "player"), measures = "Base",
  seasons = 2018, modes = c("PerGame", "Totals"),
  season_types = "Regular Season", playoff_rounds = NA, is_plus_minus = F,
  is_rank = F, is_pace_adjusted = F, outcomes = NA, locations = NA,
  months = NA, season_segments = NA, date_from = NA, date_to = NA,
  opponent_ids = NA, vs_confs = NA, vs_divisions = NA,
  game_segments = NA, periods = NA, shot_clocks = NA, last_n_games = NA,
  assign_to_environment = T, return_messages = TRUE)
}
\arguments{
\item{return_messages}{}
}
\description{
Title
}
