% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_salaries.R
\name{get_teams_yahoo_player_salary_data}
\alias{get_teams_yahoo_player_salary_data}
\title{get_teams_yahoo_player_salary_data}
\usage{
get_teams_yahoo_player_salary_data(team_names = c("Boston Celtics",
  "Brooklyn Nets", "New York Knicks"), use_all_teams = F,
  format_currency = T, nest_data = T, return_wide = F,
  return_message = T)
}
\arguments{
\item{return_wide}{}
}
\description{
get_teams_yahoo_player_salary_data
}

