% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_tables.R
\name{get_teams_year_by_year_stats}
\alias{get_teams_year_by_year_stats}
\title{NBA Year-over-year franchise history}
\usage{
get_teams_year_by_year_stats(teams = NULL, team_ids = NULL,
  all_active_teams = T, season_types = c("Regular Season"),
  modes = c("Totals"), return_message = TRUE, nest_data = F)
}
\arguments{
\item{teams}{vector of team names}

\item{team_ids}{vecor of team ids}

\item{all_active_teams}{if \code{TRUE} returns all active teams}

\item{season_types}{type of season options include \itemize{
\item Pre Season
\item Regular Season
\item Playoffs
\item All Star
}}

\item{modes}{mode of search options include \itemize{
\item PerGame
\item Totals
}}

\item{return_message}{if \code{TRUE} returns a message}

\item{nest_data}{if \code{TRUE} returns a nested data_frame}
}
\value{
a \code{data_frame}
}
\description{
NBA Year-over-year franchise history
}
\examples{
get_teams_year_by_year_stats(all_active_teams = T, modes = c("Totals"), return_message = TRUE, nest_data =F)
}
