% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_salaries.R
\name{get_yahoo_roster_salary_data}
\alias{get_yahoo_roster_salary_data}
\title{get_roster_salary_data}
\usage{

  get_yahoo_roster_salary_data(url = "http://sports.yahoo.com/nba/teams/bro/salary/",
  nest_data = T, return_wide = F, format_currency = T,
  return_message = T)
}
\arguments{
\item{return_wide}{}
}
\description{
get_roster_salary_data
}

