% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draft_combine_data.R
\name{get_year_draft_combine}
\alias{get_year_draft_combine}
\title{Get Year Draft Combines}
\usage{
get_year_draft_combine(combine_year = 2014, return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get Year Draft Combines
}

