% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_functions.R, R/teams_rosters_history.R
\name{height_in_inches}
\alias{height_in_inches}
\title{Height in Inches}
\usage{
height_in_inches(height)

height_in_inches(height)
}
\arguments{
\item{height}{}

\item{height}{}
}
\description{
Height in Inches

Get Height in Inches
}

