% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoops_hype_functions.R
\name{hoops_hype_salary_summary}
\alias{hoops_hype_salary_summary}
\title{HoopsHype NBA teams summary salaries}
\usage{
hoops_hype_salary_summary(return_wide = T, return_message = T)
}
\arguments{
\item{return_wide}{if \code{TRUE} returns wide data}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{tibble()}
}
\description{
Gets all team salary summaries from HoopsHype
}
\examples{
hoops_hype_salary_summary(spread_data = TRUE)
}
\seealso{
Other hoopshype: 
\code{\link{hoopshype_salaries}()}

Other salaries: 
\code{\link{bref_bios}()},
\code{\link{hoopshype_salaries}()},
\code{\link{nba_insider_salaries}()}
}
\concept{hoopshype}
\concept{salaries}
