% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoops_hype_functions.R
\name{hoopshype_salaries}
\alias{hoopshype_salaries}
\title{Hoopshype teams players salaries}
\usage{
hoopshype_salaries(teams = NULL, all_teams = TRUE,
  assume_player_opt_out = T, assume_team_doesnt_exercise = T,
  nest_data = F, return_message = TRUE)
}
\arguments{
\item{teams}{vector of team names}

\item{all_teams}{if \code{TRUE} returns all teams}

\item{assume_player_opt_out}{if \code{TRUE} assumes player opt out of options}

\item{assume_team_doesnt_exercise}{if \code{TRUE} assumes teams don't exercise options}

\item{nest_data}{if \code{TRUE} nests data}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a `tibble`
}
\description{
Detailed salaries for each player and
team from hoopshype
}
\examples{
library(dplyr)
df_salaries <-
hoopshype_salaries(all_teams = TRUE,
 nest_data = F, return_message = T)
 ## By Expiring Salary Type and Team
 df_salaries \%>\% group_by(slugSeason,  nameTeam, isFinalSeason) \%>\%
 summarise(expiringSalaries = sum(amountContractMil, na.rm = T) / 1000000) \%>\%
 arrange(nameTeam)
}
\seealso{
Other hoopshype: \code{\link{hoops_hype_salary_summary}}

Other salaries: \code{\link{bref_bios}},
  \code{\link{hoops_hype_salary_summary}},
  \code{\link{nba_insider_salaries}}
}
\concept{hoopshype}
\concept{salaries}
