% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_rosters_history.R
\name{nba_franchise_history}
\alias{nba_franchise_history}
\title{Get NBA franchise history}
\usage{
nba_franchise_history(return_message = T, only_active = F)
}
\arguments{
\item{return_message}{if \code{TRUE} returns message}

\item{only_active}{if \code{TRUE} only active teams}
}
\value{
a \code{tibble()}
}
\description{
Acquires history for all NBA franchises
}
\examples{
nba_franchise_history()
}
