% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injuries.R
\name{nba_injuries}
\alias{nba_injuries}
\title{NBA Injuries}
\usage{
nba_injuries(filter_returning_today = T)
}
\arguments{
\item{filter_returning_today}{if \code{TRUE} filter players returning today}
}
\value{
a `data_frame`
}
\description{
Returns information about NBA injured players
}
\examples{
nba_injuries()
}
\seealso{
Other players: \code{\link{metrics_leaders}},
  \code{\link{players_bios}},
  \code{\link{teams_players_stats}}
}
\concept{injuries}
\concept{players}
