% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_insider_functions.R
\name{nba_insider_salaries}
\alias{nba_insider_salaries}
\title{NBA team salaries}
\usage{
nba_insider_salaries(assume_player_opt_out = T,
  assume_team_doesnt_exercise = T, return_wide = F,
  return_message = T)
}
\arguments{
\item{assume_player_opt_out}{if `TRUE` assumes player opts out of a player option}

\item{assume_team_doesnt_exercise}{if `TRUE` assumes teams do not exercise team option}

\item{return_wide}{if `TRUE` spreads data}

\item{return_message}{if `TRUE` returns a message}
}
\value{
a `data_frame`
}
\description{
Gets information about NBA teams salaries
}
\examples{
nba_insider_salaries(assume_player_opt_out = T, assume_team_doesnt_exercise = T, return_message = TRUE)
}
\seealso{
Other salaries: \code{\link{bref_bios}},
  \code{\link{hoops_hype_salary_summary}},
  \code{\link{hoopshype_salaries}}
}
\concept{salaries}
