% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ish.R
\name{nba_player_ids}
\alias{nba_player_ids}
\title{Players' NBA player ids}
\usage{
nba_player_ids(players = NULL, player_ids = NULL)
}
\arguments{
\item{players}{vector of player names}

\item{player_ids}{vector of player ids}
}
\value{
a `tibble`
}
\description{
Resolve player ids for specified players
}
\examples{
nba_player_ids(players = c("Mitch Richmond", "Kyle Kuzma"))
}
\seealso{
Other ids: \code{\link{nba_teams_ids}}
}
\concept{ids}
