% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{nba_teams}
\alias{nba_teams}
\title{NBA team dictionary}
\usage{
nba_teams(join_blg = T)
}
\arguments{
\item{join_blg}{if \code{TRUE} joins BLG team features data}
}
\value{
a `data_frame`
}
\description{
Returns team dictionary
}
\examples{
nba_teams()
}
