% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{nbastats_api_parameters}
\alias{nbastats_api_parameters}
\title{NBA Stats API Parameters}
\usage{
nbastats_api_parameters(api_versions = 1:3)
}
\arguments{
\item{api_versions}{\itemize{
\item 1  - V1
\item 2 - V2
\item 3 - V3

}}
}
\value{
\code{tibble()}
}
\description{
NBA Stats API Parameters
}
\examples{
nbastats_api_parameters(1:3)
}
