% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{parse_records}
\alias{parse_records}
\title{Parse Record columns}
\usage{
parse_records(data, record_names)
}
\arguments{
\item{data}{a \code{data_frame}}

\item{record_names}{vecor of record_names}
}
\description{
Parse Record columns
}
