% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_salaries.R
\name{parse_yahoo_team_salary_data}
\alias{parse_yahoo_team_salary_data}
\title{parse_yahoo_team_salary_data}
\usage{
parse_yahoo_team_salary_data(page, return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\description{
parse_yahoo_team_salary_data
}
