% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_by_play.R
\name{play_by_play_v2}
\alias{play_by_play_v2}
\title{NBA games play-by-play v2}
\usage{
play_by_play_v2(game_ids = NULL, nest_data = FALSE,
  return_message = TRUE)
}
\arguments{
\item{game_ids}{vector of game ids}

\item{nest_data}{if \code{TRUE} nests data}

\item{return_message}{if \code{T} returns message}
}
\value{
a \code{tibble}
}
\description{
Returns play-by-play information using the playbyplayv2 endpoint
}
\examples{
play_by_play(game_ids = c(21700002, 21700003), nest_data = F, return_message = T)
}
