% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_gm.R
\name{players_agents}
\alias{players_agents}
\title{Get NBA Players Agents}
\usage{
players_agents(nest_data = F, return_message = T)
}
\arguments{
\item{return_message}{if \code{TRUE} returns player url}
}
\value{
a `tibble()`
}
\description{
Get NBA Players Agents
}
\examples{
library(dplyr)
df_players_agents <-
 players_agents()
df_players_agents \%>\%
glimpse()
}
\seealso{
Other agents: 
\code{\link{agents_players}()}
}
\concept{agents}
