% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_profile.R
\name{players_awards}
\alias{players_awards}
\title{NBA players awards}
\usage{
players_awards(players = NULL, player_ids = NULL, nest_data = F,
  return_message = TRUE)
}
\arguments{
\item{players}{vector of players}

\item{player_ids}{vector of player ids}

\item{nest_data}{if \code{TRUE} returns nested data_frame}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{data_frame}
}
\description{
Gets NBA awards for
specified players
}
\examples{
players_awards(players = c( "Charles Oakley", "Gary Melchionni"),
player_ids = c(893, 76375),
return_message = T,
 nest_data = F)
}
\seealso{
Other player: \code{\link{bref_bios}},
  \code{\link{draft_combines}},
  \code{\link{players_careers}},
  \code{\link{players_tables}}

Other awards: \code{\link{bref_awards_votes}},
  \code{\link{bref_awards}}
}
\concept{awards}
\concept{player}
