% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_career_stats.R
\name{players_careers}
\alias{players_careers}
\title{Player career stats}
\usage{
players_careers(
  players = NULL,
  player_ids = NULL,
  modes = c("PerGame", "Totals"),
  assign_to_environment = TRUE,
  add_mode_names = TRUE,
  return_message = TRUE
)
}
\arguments{
\item{players}{\code{NULL} or \code{vector} of NBA players}

\item{player_ids}{\code{vector} of NBA Player IDs}

\item{modes}{\code{vector} of items that can include \itemize{
\item Totals
\item PerGame
\item Per36
}}

\item{assign_to_environment}{if \code{TRUE} assign to each table environment
with a name starting with data}

\item{add_mode_names}{if \code{TRUE} adds the type of mode to the table data}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{tibble}
}
\description{
NBA player career statistics for
specified players and inputs
}
\examples{
players_careers(players = c("Joe Harris", "Myles Turner", "Spencer Dinwiddie"),
modes = c("Totals", "PerGame"))
}
\seealso{
Other player: 
\code{\link{bref_bios}()},
\code{\link{draft_combines}()},
\code{\link{players_awards}()},
\code{\link{players_tables}()}
}
\concept{player}
\concept{summary stats}
