% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nba_bokeh_shot_chart.R
\name{plot_nba_player_bokeh_shotchart}
\alias{plot_nba_player_bokeh_shotchart}
\title{Get's NBA player's season shot chart with specific parameters}
\usage{
plot_nba_player_bokeh_shotchart(player, year_season_end = 2015,
  plot_hex = T, author = "Alex Bresler", use_shot_zone_side = T,
  season_type = "Regular Season", shot_areas = c("Three Point", "Paint",
  "Mid-Range"), shots_type = c("Dunk", "Layup", "Jump Shot", "Fadeaway",
  "Bank", "Tip", "Hook"), vs_conference = NA, vs_division = NA,
  quarter_range = 1:12, minute_range = 0:12, against_team = NA,
  game_location = NA, game_month = NA, outcome = NA, position = NA,
  season_segment = NA, exclude_backcourt = T)
}
\arguments{
\item{exclude_backcourt}{}
}
\description{
Get's NBA player's season shot chart with specific parameters
}
\examples{
 plot_nba_player_bokeh_shotchart( player = "Thomas Robinson", exclude_backcourt = T, shots_type = c("Dunk", "Layup", "Hook", "Jump Shot"))

shot_areas = c('Paint', 'Mid-Range'), use_shot_zone_side = T, season_segment = 'Post All-Star',
year_season_end = 2015, author = "Alex Bresler")
}
