% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nba_teams_season_bokeh_shot_chart.R
\name{plot_nba_team_season_bokeh_shotchart}
\alias{plot_nba_team_season_bokeh_shotchart}
\title{Gets NBA team Bokeh Shot Chart}
\usage{
plot_nba_team_season_bokeh_shotchart(team = "Nets", year_roster = 2016,
  year_data = 2015, plot_hex = T, author = "Alex Bresler",
  use_shot_zone_side = T, season_type = "Regular Season",
  positions = c("G", "F", "C"), shots_type = c("Dunk", "Layup", "Jump Shot",
  "Fadeaway", "Bank", "Tip", "Hook"), shot_areas = c("Three Point", "Paint",
  "Mid-Range"), vs_conference = NA, vs_division = NA,
  quarter_range = 1:12, minute_range = 0:12, against_team = NA,
  game_location = NA, game_month = NA, outcome = NA,
  season_segment = NA, exclude_backcourt = T)
}
\arguments{
\item{team}{can be = #c("76ers", "Bucks", "Bulls", "Cavaliers", "Celtics", "Clippers",}

\item{year_roster}{- year of the roster}

\item{plot_hex}{- T or F, shows}

\item{season_type}{- Regular Season, Pre Season, Playoffs, All Star}

\item{positions}{- G, FA, C}

\item{shots_type}{-  ('Dunk', 'Layup', 'Jump Shot', "Fadeaway", "Bank", "Tip",'Hook')}

\item{shot_areas}{- c('Three Point', "Paint", "Mid-Range")}

\item{vs_conference}{- NA or East, West,}

\item{vs_division}{- Atlantic, Central, Northwest, Pacific, Southeast, Southwest}

\item{quarter_range}{- 0:8}

\item{minute_range}{- 0:12}

\item{against_team}{- Same as team}

\item{game_location}{- NA or Home, Road}

\item{outcome}{- NA, W, L}

\item{season_segment}{NA, Post All-Star, Pre All-Star}

\item{exclude_backcourt}{}
}
\description{
Gets NBA team Bokeh Shot Chart
}
\examples{
plot_nba_team_season_bokeh_shotchart(team = "Nets", plot_hex = F, quarter_range = c(3,4))
}
