% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_per_minute}
\alias{scale_per_minute}
\title{Summarise data per minute}
\usage{
scale_per_minute(data, scale_columns = NULL)
}
\arguments{
\item{data}{a data frame}

\item{scale_columns}{vector of columns to scale}
}
\description{
Summarise data per minute
}
