% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons_rosters.R
\name{seasons_rosters}
\alias{seasons_rosters}
\title{NBA teams seasons rosters}
\usage{
seasons_rosters(seasons = NULL, return_message = TRUE, nest_data = F)
}
\arguments{
\item{seasons}{vector of seasons}

\item{return_message}{if \code{TRUE} returns a message}

\item{nest_data}{if \code{TRUE} nests data}
}
\value{
a `date_frame`
}
\description{
Rosters for each team of specified seasons
}
\examples{
library(nbastatR)
library(dplyr)
df_rosters <- seasons_rosters(2015:2018)

### Mean Age by Season and Team
df_rosters \%>\%
group_by(slugSeason, slugTeam) \%>\%
summarise(ageMean = mean(agePlayer)) \%>\%
arrange(ageMean) \%>\%
ungroup()
}
\seealso{
Other teams: \code{\link{drafts}},
  \code{\link{franchise_leaders}},
  \code{\link{teams_players_stats}},
  \code{\link{teams_rankings}}

Other rosters: \code{\link{team_season_roster}},
  \code{\link{teams_rosters}}
}
\concept{rosters}
\concept{teams}
