% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summer_league.R
\name{sl_players}
\alias{sl_players}
\title{Summer League Players}
\usage{
sl_players(year = 2018, assign_to_environment = T)
}
\arguments{
\item{year}{summer league year}

\item{assign_to_environment}{if \code{true assigns to environment}}
}
\value{
a nested \code{tibble}
}
\description{
Summer League Players
}
\examples{
sl_players()
}
