% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summer_league.R
\name{sl_teams}
\alias{sl_teams}
\title{Summer League Teams}
\usage{
sl_teams(assign_to_environment = T)
}
\arguments{
\item{assign_to_environment}{if \code{TRUE} assings to environment}
}
\value{
a nested \code{tibble}
}
\description{
Summer League Teams
}
\examples{
sl_teams()
}
