% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spread_data}
\alias{spread_data}
\title{Spread gathered data frame}
\usage{
spread_data(data, variable_name = "item", value_name = "value",
  perserve_order = TRUE, unite_columns = NULL, seperate_columns = NULL)
}
\arguments{
\item{data}{a \code{data_frame}}

\item{variable_name}{name of variable vector}

\item{value_name}{name of value vector}

\item{perserve_order}{if \code{TRUE} perserve order}

\item{seperate_columns}{}
}
\value{
a \code{data_frame}
}
\description{
Spread gathered data frame
}
