% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{synergy}
\alias{synergy}
\title{Get Synergy data for specified season}
\usage{
synergy(
  seasons = 2016:2018,
  result_types = c("player", "team"),
  season_types = c("Regular Season"),
  set_types = c("offensive", "defensive"),
  categories = c("Transition", "Isolation", "PRBallHandler", "PRRollman", "Postup",
    "Spotup", "Handoff", "Cut", "OffScreen", "OffRebound", "Misc"),
  results = 500,
  assign_to_environment = TRUE,
  return_wide = F,
  nest_data = F,
  return_message = TRUE
)
}
\arguments{
\item{seasons}{vector of seasons from 2016 onward}

\item{result_types}{result type \itemize{
\item team
\item player
}}

\item{season_types}{type of season play \itemize{
\item Playoffs
\item Regular Season
}}

\item{set_types}{set type \itemize{
\item offensive
\item defensive
}}

\item{categories}{vector of synergy categories options include: \itemize{
\item Transition
\item Isolation
\item PRBallHandler
\item PRRollman
\item Postup
\item Spotup
\item Handoff
\item Cut
\item OffScreen
\item OffRebound
\item Misc
}}

\item{results}{number of results}

\item{assign_to_environment}{if \code{TRUE} assigns table to environment}

\item{return_wide}{if \code{return_wide} returns a spread \code{tibble}}

\item{return_message}{}
}
\value{
a \code{tibble}
}
\description{
Get Synergy data for specified result type,
category, season type and set
}
\examples{
synergy(seasons = 2019, result_types = c("player", "team"), season_types = c("Regular Season"), set_types = c("offensive", "defensive"), categories = c("Transition", "Isolation", "PRBallHandler", "PRRollman", "Postup",  "Spotup", "Handoff", "Cut", "OffScreen", "OffRebound", "Misc"), results = 500, assign_to_environment = TRUE, return_wide = F, return_message = TRUE)
}
