% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_rosters_history.R
\name{team_season_roster}
\alias{team_season_roster}
\title{Team roster}
\usage{
team_season_roster(team = "Denver Nuggets", season = 2015,
  return_message = T)
}
\arguments{
\item{team}{team name}

\item{season}{season vector}

\item{return_message}{}
}
\value{
a `tibble`
}
\description{
Gets teams roster from a specified seasons
}
\examples{
team_season_roster(team = "Denver Nuggets", season = 1991)
}
\seealso{
Other rosters: \code{\link{seasons_rosters}},
  \code{\link{teams_rosters}}
}
\concept{rosters}
