% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_tables.R
\name{teams_annual_stats}
\alias{teams_annual_stats}
\title{NBA teams yearly performance}
\usage{
teams_annual_stats(teams = NULL, team_ids = NULL,
  all_active_teams = T, season_types = c("Regular Season"),
  modes = c("Totals"), return_message = TRUE, nest_data = F)
}
\arguments{
\item{teams}{vector of team names}

\item{team_ids}{vector of team ids}

\item{all_active_teams}{if \code{TRUE} returns all active teams}

\item{season_types}{type of season options include \itemize{
\item Pre Season
\item Regular Season
\item Playoffs
\item All Star
}}

\item{modes}{mode of search options include \itemize{
\item PerGame
\item Totals
}}

\item{return_message}{if \code{TRUE} returns a message}

\item{nest_data}{if \code{TRUE} returns a nested tibble}
}
\value{
a \code{tibble}
}
\description{
NBA teams yearly performance
}
\examples{
teams_annual_stats(all_active_teams = T,
modes = c("Totals"),
return_message = TRUE,
 nest_data =F)
}
