% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_rosters_history.R
\name{teams_coaches}
\alias{teams_coaches}
\title{Seasons coaching staffs}
\usage{
teams_coaches(seasons = 1990:2018, nest_data = F, return_message = T)
}
\arguments{
\item{seasons}{vector of seasons}

\item{nest_data}{if \code{TRUE} nests data}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a `data_frame`
}
\description{
Gets coaching staffs for specified seasons
}
\examples{
teams_coaches(2018:2019)
}
