% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_teams.R
\name{teams_players_stats}
\alias{teams_players_stats}
\title{NBA teams and player statistics}
\usage{
teams_players_stats(seasons = 2018, types = c("player", "team"),
  tables = c("defense"), season_types = "Regular Season",
  measures = "Base", modes = "PerGame", defenses = "Overall",
  is_plus_minus = F, is_pace_adjusted = F, periods = 0,
  is_rank = F, game_segments = NA, divisions_against = NA,
  conferences_against = NA, date_from = NA, date_to = NA,
  last_n_games = 0, locations = NA, months = 0,
  season_segments = NA, opponents = NA, countries = NA,
  weights = NA, outcomes = NA, playoff_rounds = 0,
  players_experience = NA, players_positions = NA, colleges = NA,
  draft_picks = NA, draft_years = NA, game_scopes = NA,
  heights = NA, shot_clock_ranges = NA,
  clutch_times = "Last 5 Minutes", ahead_or_behind = "Ahead or Behind",
  general_ranges = "Overall", dribble_ranges = "0 Dribbles",
  shot_distance_ranges = "By Zone", touch_time_ranges = NA,
  closest_defender_ranges = NA, point_diffs = 5, starters_bench = NA,
  assign_to_environment = TRUE, add_mode_names = T,
  return_message = TRUE)
}
\arguments{
\item{seasons}{vector of seasons}

\item{types}{type of data options include \itemize{
\item player
\item team
}}

\item{tables}{type of table options include \itemize{
\item clutch
\item defense
\item general
\item hustle
\item shot locations
\item shots
\item splits
}}

\item{season_types}{vector of season types options \itemize{
\item Pre Season
\item Regular Season
\item Playoffs
\item All Star
}}

\item{measures}{vector of measures options \itemize{
\item Base
\item Advanced
\item Defense
\item Four Factors
\item Misc
\item Opponent
\item Scoring
\item Usage
}}

\item{modes}{vector of modes options \itemize{
\item Totals
\item MinutesPer
\item PerGame
\item Per48
\item Per40
\item Per36
\item PerMinute
\item PerPossession
\item PerPlay
\item Per100Possessions
\item Per100Plays
}}

\item{defenses}{defense options include \itemize{
\item     Overall
\item 3 Pointers
\item 2 Pointers
\item Less Than 6Ft
\item Less Than 10Ft
\item Greater Than 15Ft
}}

\item{is_plus_minus}{if `TRUE` uses plus minus}

\item{is_pace_adjusted}{if `TRUE` is pace adjusted}

\item{periods}{vector of periods}

\item{is_rank}{if `TRUE` returns rank}

\item{game_segments}{vector of game segments options \itemize{
\item NA - All
\item First Half
\item Second Half
\item Overtime
}}

\item{divisions_against}{vector of seasons options \itemize{
\item NA - ALL
\item Atlantic
\item Central
\item Northwest
\item Pacific
\item Southeast
\item Southwest
\item East
\item West
}}

\item{conferences_against}{vector of conferences against options \itemize{
\item NA - ALL
\item East
\item West
}}

\item{date_from}{dates from}

\item{date_to}{dates to}

\item{last_n_games}{vector games}

\item{locations}{vector of locations}

\item{months}{vector of months 0:12}

\item{season_segments}{vector of seasons segments options \itemize{
\item NA
\item Pre All-Star
\item Post All-Star
}}

\item{opponents}{vector of opponent ids}

\item{countries}{vector of countries}

\item{weights}{vector of weights options \itemize{
\item NA
\item LT 200
\item GT 200
\item LT 225
\item GT 225
\item LT 250
\item GT 250
\item LT 275
\item GT 275
\item LT 300
\item GT 300
}}

\item{outcomes}{vector of outcomes options \itemize{
\item NA - all
\item W
\item L
}}

\item{playoff_rounds}{vector of playoff rounds options \code{0:4}}

\item{players_experience}{vector of experience options \itemize{
\item NA
\item Rookie
\item Sophomore
\item Veteran
}}

\item{players_positions}{vector of player positions options \itemize{
\item NA
\item C
\item F
\item G
}}

\item{colleges}{vector of colleges}

\item{draft_picks}{vector of draft picks options \itemize{
\item NA
\item     1st Round
\item 2nd Round
\item 1st Pick
\item Lottery Pick
\item Top 5 Pick
\item Top 10 Pick
\item Top 15 Pick
\item Top 20 Pick
\item Top 25 Pick
\item Picks 11 Thru 20
\item Picks 21 Thru 30
\item Undrafted
}}

\item{draft_years}{numeric vector vector of draft years}

\item{game_scopes}{vector game scopes options \itemize{
\item NA
\item Last 10
\item Yesterday
}}

\item{heights}{vector of heights options \itemize{
\item NA
\item     LT 6-0
\item GT 6-0
\item LT 6-4
\item GT 6-4
\item LT 6-7
\item GT 6-7
\item LT 6-10
\item GT 6-10
\item LT 7-0
\item GT 7-0
}}

\item{shot_clock_ranges}{character vector of shot clock ranges options \itemize{
\item NA
\item 24-22
\item 22-18 Very Early
\item 18-15 Early
\item 15-7 Average
\item 7-4 Late
\item 4-0 Very Late
\item ShotClock Off
}}

\item{clutch_times}{clutch options options \itemize{
\item NA
\item Last 5 Minutes
\item Last 4 Minutes
\item Last 3 Minutes
\item Last 2 Minutes
\item Last 1 Minute
\item Last 30 Seconds
\item Last 10 Seconds
}}

\item{ahead_or_behind}{ahead of behind type options \itemize{
\item Ahead or Behind
\item Behind or Tied
\item Ahead or Tied
}}

\item{general_ranges}{general shop type ranges options include \itemize{
\item Overall
\item Catch and Shoot
\item Less Than 10 ft
\item Pullups
}}

\item{dribble_ranges}{range of dribbles options include \itemize{
\item 0 Dribbles
\item 1 Dribble
\item 2 Dribbles
\item 3-6 Dribbles
\item 7+ Dribbles
}}

\item{shot_distance_ranges}{shot distance}

\item{touch_time_ranges}{touch time range options include \itemize{
\item Touch < 2 Seconds
\item Touch 2-6 Seconds
\item Touch 6+ Seconds

}}

\item{closest_defender_ranges}{closest defender range options include \itemize{
\item NA
\item 0-2 Feet - Very Tight
\item 2-4 Feet - Tight
\item 4-6 Feet - Open
\item 6+ Feet - Wide Open
}}

\item{point_diffs}{numeric vector between \code{1:5}}

\item{starters_bench}{vector of starter type options \itemize{
\item NA
\item Bench
\item Starters
}}

\item{assign_to_environment}{if `TRUE` assigns tables to environment}

\item{add_mode_names}{if `TRUE` adds mode names}

\item{return_message}{if `TRUE` returns a message}
}
\value{
a `tibble`
}
\description{
Gets NBA summary statistics tables
}
\examples{
teams_players_stats(seasons = 2018, types = c("player", "team"),
 modes = c("PerGame", "Totals"),
 tables = c("general", "defense", "clutch", "hustle", "shots", "shot locations"))
 )
}
\seealso{
Other players: \code{\link{metrics_leaders}},
  \code{\link{players_bios}}

Other teams: \code{\link{drafts}},
  \code{\link{franchise_leaders}},
  \code{\link{seasons_rosters}},
  \code{\link{teams_rankings}}
}
\concept{players}
\concept{summary statistics}
\concept{teams}
