% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_rankings.R
\name{teams_rankings}
\alias{teams_rankings}
\title{NBA teams rankings}
\usage{
teams_rankings(seasons = NULL, nest_data = F, return_message = T)
}
\arguments{
\item{seasons}{vector of numeric seasons}

\item{nest_data}{if `TRUE` nests data by season}

\item{return_message}{`TRUE` returns a message}
}
\value{
a `data_frame`
}
\description{
Team rankings for specified season
}
\examples{
teams_rankings(seasons = 2019)
}
\seealso{
Other teams: \code{\link{drafts}},
  \code{\link{franchise_leaders}},
  \code{\link{seasons_rosters}},
  \code{\link{teams_players_stats}}
}
\concept{rankings}
\concept{teams}
