% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{teams_seasons_info}
\alias{teams_seasons_info}
\title{NBA teams seasons information}
\usage{
teams_seasons_info(
  teams = NULL,
  team_ids = NULL,
  all_active_teams = T,
  seasons = 2019,
  season_types = "Regular Season",
  nest_data = F,
  return_message = T
)
}
\arguments{
\item{teams}{vector of team names}

\item{team_ids}{vector of team ids}

\item{all_active_teams}{if \code{TRUE} returns all active teams}

\item{seasons}{vector of seasons}

\item{season_types}{type of season options include \itemize{
\item Regular Season
\item Playoffs
\item Pre Season
}}

\item{nest_data}{if `TRUE` nests data}

\item{return_message}{if `TRUE` returns a message}
}
\value{
a `tibble`
}
\description{
Acquires information for a teams season
}
\examples{
teams_seasons_info(teams = "Brooklyn Nets", seasons = c(1984, 1990, 1995, 2018), season_types = "Regular Season")
}
