% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{teams_shots}
\alias{teams_shots}
\title{Get teams seasons shot charts}
\usage{
teams_shots(teams = NULL, team_ids = NULL, all_active_teams = F,
  season_types = "Regular Season", seasons = 2018, measures = "FGA",
  periods = 0, months = 0, date_from = NA, date_to = NA,
  nest_data = F, return_message = T)
}
\arguments{
\item{teams}{vector of team names}

\item{team_ids}{vector of team ids}

\item{all_active_teams}{if `TRUE` returns all active teams}

\item{season_types}{vector of season types options \itemize{
\item Pre Season
\item Regular Season
\item Playoffs
\item All Star
}}

\item{seasons}{vector of seasons}

\item{measures}{vector of measures}

\item{periods}{vector of periods}

\item{months}{vector of months}

\item{date_from}{date from}

\item{date_to}{date to}

\item{nest_data}{if `TRUE` nests data}

\item{return_message}{if `TRUE` returns a message}
}
\value{
a \code{data_frame}
}
\description{
Get teams seasons shot charts
}
\examples{
teams_shots(teams = "Brooklyn Nets",
seasons = 2018)
}
