% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{transactions}
\alias{transactions}
\title{NBA transactions since 2012}
\usage{
transactions(include_history = T)
}
\value{
a `tibble`
}
\description{
Acquires all NBA transactions since 2012
}
\examples{
transactions()
}
\concept{transactions}
