% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{validate_nba_player_photos}
\alias{validate_nba_player_photos}
\title{Validate NBA Player photos}
\usage{
validate_nba_player_photos(sleep_time = 2)
}
\arguments{
\item{sleep_time}{if not \code{NULL} sleep time}
}
\value{
\code{tibble}
}
\description{
Validate NBA Player photos
}
