% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_som.R
\name{visualize_similar_players}
\alias{visualize_similar_players}
\title{Visualize Similar Players}
\usage{
visualize_similar_players(data, names = c("Russell Westbrook",
  "Bradley Beal"), group_distance = 1, use_som_group = TRUE,
  column_order = c("nameSearch", "clusterSOM", "groupSOM", "namePlayer"),
  use_radial_network = TRUE, title = "Similar Players")
}
\arguments{
\item{title}{}
}
\description{
Visualize Similar Players
}
