% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_som.R
\name{visualize_topology_network}
\alias{visualize_topology_network}
\title{Plot Topology Network}
\usage{
visualize_topology_network(data, use_name_player_id_column = FALSE,
  group_node = "SOM", include_gng_neighbors = TRUE,
  title = "NBA Player Topology: 2015-16 Season, 250 Minute Minimum -- Alex Bresler",
  subtitle = "data normalized per minute, scaled to mean zero", footer = "")
}
\arguments{
\item{data}{a \code{data_frame}}

\item{group_node}{Grouping node \itemize{
\item \code{SOM}: Self Organizing Map
\item \code{GNG, Neural Gas}: Neural Gas Node
}}

\item{title}{Plot Title}

\item{subtitle}{Plot subtitle}

\item{footer}{Plot footer}

\item{name_player_id_column}{\code if \code{TRUE} uses player season id instead of name for nodes}
}
\description{
Plot Topology Network
}
