% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{widen_bref_data}
\alias{widen_bref_data}
\title{Widens basketball reference table data}
\usage{
widen_bref_data(data)
}
\arguments{
\item{data}{}
}
\value{
a `data_frame`
}
\description{
Widens basketball reference table data
}
