% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_by_play.R
\name{win_probability}
\alias{win_probability}
\title{NBA games win probabilities}
\usage{
win_probability(game_ids = c(21700002, 21700003), nest_data = FALSE,
  filter_non_plays = FALSE, return_message = TRUE)
}
\arguments{
\item{game_ids}{vector of game ids}

\item{nest_data}{if \code{TRUE} nests data}

\item{filter_non_plays}{if \code{TRUE} filters out non plays}

\item{return_message}{if \code{T} returns message}
}
\value{
a \code{tibble}

a `tibble`
}
\description{
Gets nba in-game win probabilities
for specified game ids
}
\examples{
win_probability(game_ids = c(21700002, 21700005), filter_non_plays = T,
nest_data = FALSE,
return_message = TRUE)
}
\seealso{
Other game: \code{\link{box_scores}},
  \code{\link{fanduel_summary}}, \code{\link{game_logs}}

Other season: \code{\link{fanduel_summary}}
}
\concept{game}
\concept{season}
