% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{assign_bref_data}
\alias{assign_bref_data}
\title{Assign nested BREF data to environment}
\usage{
assign_bref_data(data, type = "Players", widen_data = TRUE,
  join_data = TRUE, nest_data = FALSE, assign_to_environment = TRUE)
}
\arguments{
\item{data}{a \code{data_frame} of tables}

\item{type}{type of BREF data}

\item{widen_data}{if \code{TRUE} widens data}

\item{join_data}{if \code{TRUE} joins tables}

\item{nest_data}{if \code{TRUE} nests data}

\item{assign_to_environment}{if \code{TRUE} assigns data to your environment}
}
\description{
Assign nested BREF data to environment
}
