% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ish.R
\name{assign_nba_players}
\alias{assign_nba_players}
\title{Assign NBA player dictionary to environment}
\usage{
assign_nba_players()
}
\description{
Assign NBA player dictionary to environment
}
\examples{
assign_nba_players()
}
