% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ish.R
\name{assign_nba_teams}
\alias{assign_nba_teams}
\title{Assign NBA teams to environment}
\usage{
assign_nba_teams()
}
\value{
a `tibble`
}
\description{
Assign NBA teams to environment
}
\examples{
assign_nba_teams()
}
\seealso{
Other dictionary: 
\code{\link{assign_nba_players}()},
\code{\link{dictionary_bref_coaches}()},
\code{\link{dictionary_bref_players}()},
\code{\link{dictionary_bref_teams}()}
}
\concept{NBA teams}
\concept{dictionary}
