% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{dictionary_bref_coaches}
\alias{dictionary_bref_coaches}
\title{Basketball Reference coach dictionary}
\usage{
dictionary_bref_coaches()
}
\value{
a `data_frame`
}
\description{
Basketball Reference coach dictionary
}
\examples{
dictionary_bref_coaches()
}
\seealso{
Other dictionary: \code{\link{assign_nba_players}},
  \code{\link{assign_nba_teams}},
  \code{\link{dictionary_bref_players}},
  \code{\link{dictionary_bref_teams}}

Other BREF: \code{\link{bref_awards_votes}},
  \code{\link{bref_awards}},
  \code{\link{bref_players_seasons}},
  \code{\link{bref_teams_seasons}},
  \code{\link{dictionary_bref_players}},
  \code{\link{dictionary_bref_teams}},
  \code{\link{get_bref_players_bios}}

Other Coaching: \code{\link{get_coaching_staffs}}
}
\concept{BREF}
\concept{Coaching}
\concept{dictionary}
