% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_insider_functions.R
\name{get_all_nba_teams_salaries}
\alias{get_all_nba_teams_salaries}
\title{NBA team salaries}
\usage{
get_all_nba_teams_salaries(assume_player_opt_out = T,
  assume_team_doesnt_exercise = T, spread_data = F, return_message = T)
}
\arguments{
\item{assume_player_opt_out}{if `TRUE` assumes player opts out of a player option}

\item{assume_team_doesnt_exercise}{if `TRUE` assumes teams do not exercise team option}

\item{spread_data}{if `TRUE` spreads data}

\item{return_message}{if `TRUE` returns a message}
}
\value{
a `data_frame`
}
\description{
Gets information about NBA teams salaries
}
\examples{
get_all_nba_teams_salaries(assume_player_opt_out = T, assume_team_doesnt_exercise = T, return_message = TRUE)
}
\seealso{
Other salaries: \code{\link{get_bref_players_bios}},
  \code{\link{get_hoops_hype_teams_salary_summary}},
  \code{\link{get_nba_team_salaries}},
  \code{\link{get_teams_hoopshype_salaries}}
}
