% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{get_beyond_the_numbers_articles}
\alias{get_beyond_the_numbers_articles}
\title{Get Beyond the Numbers articles}
\usage{
get_beyond_the_numbers_articles(count_articles = 50)
}
\arguments{
\item{count_articles}{numeric vector of counts}
}
\value{
a \code{data_frame}
}
\description{
Get Beyond the Numbers articles
}
\examples{

get_beyond_the_numbers_articles(10)
}
