% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{get_bref_all_nba_teams}
\alias{get_bref_all_nba_teams}
\title{All NBA teams}
\usage{
get_bref_all_nba_teams(only_nba = T, return_message = T)
}
\arguments{
\item{only_nba}{if `TRUE` returns only NBA all NBA teams}

\item{return_message}{if `TRUE` returns a message}
}
\value{
a `data_frame`
}
\description{
All NBA teams
}
\examples{
get_bref_all_nba_teams()
}
\seealso{
Other awards: \code{\link{get_bref_awards}},
  \code{\link{get_bref_seasons_award_votes}},
  \code{\link{get_players_awards}}
}
