% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{get_bref_coach_dictionary}
\alias{get_bref_coach_dictionary}
\title{Basketball Reference Coach Dictionary}
\usage{
get_bref_coach_dictionary()
}
\description{
Basketball Reference Coach Dictionary
}
\examples{
get_bref_coach_dictionary()
}
