% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref_bio.R
\name{get_bref_players_bios}
\alias{get_bref_players_bios}
\title{Basketball-Reference players bios}
\usage{
get_bref_players_bios(players = NULL, player_ids = NULL,
  assign_to_environment = T, return_message = T)
}
\arguments{
\item{players}{vector of player names}

\item{player_ids}{vector of basketball reference player ids}

\item{assign_to_environment}{if \code{TRUE} assigns each table to environment}

\item{return_message}{if \code{TRUE} returns}
}
\value{
a \code{data_frame}
}
\description{
Includes player bios, salaries, contracts and transactions
}
\examples{
get_bref_players_bios( players = c("Jarrett Allen", "Mitch Richmond", "Michael Adams"),  player_ids = NULL, assign_to_environment = TRUE)
}
