% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{get_coaching_staffs}
\alias{get_coaching_staffs}
\title{NBA active coaching staffs}
\usage{
get_coaching_staffs()
}
\value{
a \code{data_frames}
}
\description{
Gets active coaching staff information for all
NBA teams
}
\examples{
get_coaching_staffs()
}
