% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_express.R
\name{get_data_draft_express_measurements}
\alias{get_data_draft_express_measurements}
\title{Draft Express Measurements}
\usage{
get_data_draft_express_measurements(return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Draft Express Measurements
}
