% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{get_days_nba_scores}
\alias{get_days_nba_scores}
\title{Get NBA Dates' NBA Scores}
\usage{
get_days_nba_scores(game_dates = NULL, day_offset = 0,
  include_standings = F, assign_to_environment = T, return_message = TRUE)
}
\arguments{
\item{game_dates}{vector of dates}

\item{day_offset}{day offset}

\item{include_standings}{if \code{TRUE} includes standings as of the date}

\item{assign_to_environment}{if \code{TRUE} assigns each table to environment with name containing dataScore}

\item{return_message}{if \code{TRUE} returns a message}
}
\description{
Returns nba score data for a given date
}
\examples{
get_days_nba_scores(game_dates = "2017-12-31", include_standings = F, return_message = T)
}
