% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_by_play.R
\name{get_games_pbp_win_probablity}
\alias{get_games_pbp_win_probablity}
\title{NBA games win probabilities}
\usage{
get_games_pbp_win_probablity(game_ids = c(21700002, 21700003),
  nest_data = FALSE, filter_non_plays = FALSE, return_message = TRUE)
}
\arguments{
\item{game_ids}{vector of game ids}

\item{nest_data}{if \code{TRUE} nests data}

\item{filter_non_plays}{if \code{TRUE} filters out non plays}

\item{return_message}{if \code{T} returns message}
}
\value{
a \code{data_frame}

a `data_frame`
}
\description{
Gets nba in-game win probabilities
for specified game ids
}
\examples{
get_games_pbp_win_probablity(game_ids = c(21700002, 21700005), filter_non_plays = T,
nest_data = FALSE,
return_message = TRUE)
}
\seealso{
Other game: \code{\link{get_game_logs}},
  \code{\link{get_games_box_scores}},
  \code{\link{get_games_fanduel_summary}}

Other season: \code{\link{get_games_fanduel_summary}}
}
\concept{game}
\concept{season}
