% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoops_hype_functions.R
\name{get_hoops_hype_teams_salary_summary}
\alias{get_hoops_hype_teams_salary_summary}
\title{HoopsHype NBA teams summary salaries}
\usage{
get_hoops_hype_teams_salary_summary(spread_data = F, return_message = T)
}
\arguments{
\item{spread_data}{if \code{TRUE} spreads data}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{data_frame()}
}
\description{
Gets all team salary summaries from HoopsHype
}
\examples{
get_hoops_hype_teams_salary_summary(spread_data = TRUE)
}
\seealso{
Other hoopshype: \code{\link{get_teams_hoopshype_salaries}}

Other salaries: \code{\link{get_all_nba_teams_salaries}},
  \code{\link{get_bref_players_bios}},
  \code{\link{get_nba_team_salaries}},
  \code{\link{get_teams_hoopshype_salaries}}
}
\concept{hoopshype}
\concept{salaries}
