% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{get_nba_current_season_schedule}
\alias{get_nba_current_season_schedule}
\title{Get Current Season Schedule}
\usage{
get_nba_current_season_schedule()
}
\description{
Get Current Season Schedule
}
\examples{
get_nba_current_season_schedule()
}
