% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ish.R
\name{get_nba_players_ids}
\alias{get_nba_players_ids}
\title{Players' NBA player ids}
\usage{
get_nba_players_ids(players = NULL, player_ids = NULL)
}
\arguments{
\item{players}{vector of player names}

\item{player_ids}{vector of player ids}
}
\value{
a `data_frame`
}
\description{
Resolve player ids for specified players
}
\examples{
get_nba_players_ids(players = c("Mitch Richmond", "Kyle Kuzma"))
}
\seealso{
Other ids: \code{\link{get_nba_teams_ids}}
}
