% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_profile.R
\name{get_nba_seasons_players}
\alias{get_nba_seasons_players}
\title{Get NBA Seasons Players}
\usage{
get_nba_seasons_players(years = 1960:2015, only_on_roster = F,
  message = T)
}
\description{
Get NBA Seasons Players
}

