% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{get_nba_stats_api_items}
\alias{get_nba_stats_api_items}
\title{Get NBA Stats API items, players and teams}
\usage{
get_nba_stats_api_items()
}
\arguments{
\item{include_tables}{}
}
\description{
Get NBA Stats API items, players and teams
}
\examples{
get_nba_stats_api_items()
}
