% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats.R
\name{get_nba_stats_api_items}
\alias{get_nba_stats_api_items}
\title{NBA stats API parameters, teams and items}
\usage{
get_nba_stats_api_items()
}
\value{
a `data_frame`
}
\description{
NBA stats API parameters, teams and items
}
\examples{
get_nba_stats_api_items()
}
