% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_rankings.R
\name{get_nba_team_current_rankings}
\alias{get_nba_team_current_rankings}
\title{NBA Team Rankings}
\usage{
get_nba_team_current_rankings()
}
\value{
a `data_frame`
}
\description{
Gets most recent NBA team rankings
}
\examples{
get_nba_team_current_rankings()
}
