% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_insider_functions.R
\name{get_nba_team_salaries}
\alias{get_nba_team_salaries}
\title{Get NBA team salaries}
\usage{
get_nba_team_salaries(team_name = "Brooklyn Nets", team_slug = NA,
  assume_player_opt_out = T, assume_team_doesnt_exercise = T,
  spread_data = FALSE, return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get NBA team salaries
}
