% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ish.R
\name{get_nba_teams_ids}
\alias{get_nba_teams_ids}
\title{Get NBA Team IDs}
\usage{
get_nba_teams_ids(teams = NULL, team_ids = NULL)
}
\arguments{
\item{teams}{vector of team names}

\item{team_ids}{vector of team ids}
}
\description{
Get NBA Team IDs
}
\examples{
get_nba_teams_ids(teams = c("Brooklyn Nets", "Denver Nuggets"))
}
