% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockdrafts.R
\name{get_nbadraftnet_mock_drafts}
\alias{get_nbadraftnet_mock_drafts}
\title{Get NBADraft.net mock drafts}
\usage{
get_nbadraftnet_mock_drafts(years = 2009:2018, merge_nba_data = T,
  nest_data = F, return_message = T)
}
\arguments{
\item{years}{vector of draft years}

\item{merge_nba_data}{if \code{TRUE} merges NBA player data}

\item{nest_data}{if \code{TRUE} nests data by draft}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{data_frame()}
}
\description{
Get NBADraft.net mock drafts
}
\examples{
get_nbadraftnet_mock_drafts(years = 2012:2017, merge_nba_data = TRUE, nest_data = F, return_message = T)
}
