% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_gm.R
\name{get_players_agents}
\alias{get_players_agents}
\title{Get NBA Players Agents}
\usage{
get_players_agents(nest_data = F, return_message = T)
}
\arguments{
\item{return_message}{if \code{TRUE} returns player url}
}
\value{
a `data_frame()`
}
\description{
Get NBA Players Agents
}
\examples{
library(dplyr)
df_players_agents <-
 get_players_agents()
df_players_agents \%>\%
glimpse()
}
\seealso{
Other agents: \code{\link{get_agents_players}}
}
