% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_profile.R
\name{get_players_awards}
\alias{get_players_awards}
\title{NBA players awards}
\usage{
get_players_awards(players = NULL, player_ids = NULL, nest_data = F,
  return_message = TRUE)
}
\arguments{
\item{players}{vector of players}

\item{player_ids}{vector of player ids}

\item{nest_data}{if \code{TRUE} returns nested data_frame}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{data_frame}
}
\description{
Gets NBA awards for
specified players
}
\examples{
get_players_awards(players = c( "Charles Oakley", "Gary Melchionni"),
player_ids = c(893, 76375),
return_message = T,
 nest_data = F)
}
\seealso{
Other player: \code{\link{get_bref_players_bios}},
  \code{\link{get_players_career_stats}},
  \code{\link{get_players_tables_data}},
  \code{\link{get_years_draft_combines}}

Other awards: \code{\link{get_bref_all_nba_teams}},
  \code{\link{get_bref_awards}},
  \code{\link{get_bref_seasons_award_votes}}
}
