% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_profile.R
\name{get_players_awards}
\alias{get_players_awards}
\title{NBA players awatds}
\usage{
get_players_awards(players = NULL, player_ids = NULL, nest_data = F,
  return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
NBA players awatds
}
\examples{
get_players_awards(players = c( "Charles Oakley", "Gary Melchionni"), player_ids = c(893, 76375), return_message = T, nest_data = F)
}
