% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_profile.R
\name{get_players_bios}
\alias{get_players_bios}
\title{Get Player bios}
\usage{
get_players_bios(players = c("Carmelo Anthony", "Joe Johnson"),
  player_ids = NULL, return_message = TRUE)
}
\arguments{
\item{players}{vector of players}

\item{player_ids}{vector of player ids}

\item{return_message}{if \code{TRUE} returns a message}
}
\description{
Seems to have ended after 2013-14 season
}
