% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_team_dash.R
\name{get_players_seasons_summary_stats}
\alias{get_players_seasons_summary_stats}
\title{Get Players Seasons Summary Statistics}
\usage{
get_players_seasons_summary_stats(seasons = 2018,
  season_types = "Regular Season", measures = "Base", modes = "PerGame",
  is_plus_minus = F, is_pace_adjusted = F, periods = 0, is_rank = F,
  game_segments = NA, divisions_against = NA, conferences_against = NA,
  date_from = NA, date_to = NA, last_n_games = 0, locations = NA,
  months = 0, season_segments = NA, opponents = NA, countries = NA,
  weights = NA, outcomes = NA, playoff_rounds = 0,
  players_experience = NA, players_positions = NA, colleges = NA,
  draft_picks = NA, draft_years = NA, game_scopes = NA, heights = NA,
  shot_clock_ranges = NA, starters_bench = NA,
  assign_to_environment = TRUE, add_mode_names = T, return_message = TRUE)
}
\arguments{
\item{weights}{}
}
\description{
Get Players Seasons Summary Statistics
}
\examples{
get_players_seasons_summary_stats(seasons = 2016:2018)
}
