% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{get_players_tables_data}
\alias{get_players_tables_data}
\title{NBA players table data}
\usage{
get_players_tables_data(players = NULL, player_ids = NULL,
  tables = c("year over year", "passes", "game splits"), measures = "Base",
  seasons = 2018, modes = c("PerGame", "Totals"),
  season_types = "Regular Season", playoff_rounds = NA, is_plus_minus = F,
  n_games = 20, is_rank = F, is_pace_adjusted = F, outcomes = NA,
  locations = NA, months = NA, season_segments = NA, date_from = NA,
  date_to = NA, opponent_ids = NA, vs_confs = NA, vs_divisions = NA,
  game_segments = NA, periods = NA, shot_clocks = NA, last_n_games = NA,
  assign_to_environment = TRUE, return_message = TRUE)
}
\arguments{
\item{players}{vector of player names}

\item{player_ids}{vector of player ids}

\item{tables}{vector of tables options \itemize{
\item passes
\item clutch
\item game splits
\item general splits
\item opponent
\item next n games
\item player on off details
\item defense
\item game logs
\item rebounding
\item shot chart detail
\item shots
\item year over year
\item fantasy profile
}}

\item{measures}{vector of measure types options include \itemize{
\item Base
\item Advanced
\item Misc
\item Scoring
\item Four Factors
\item Opponent
\item Usage
\item Defense
}}

\item{seasons}{vector of seasons}

\item{modes}{vector of modes options include \itemize{
\item PerGame
\item Totals
\item MinutesPer
\item Per48
\item Per40
\item Per36
\item PerMinute
\item PerPossession
\item PerPlay
\item Per100Possessions
\item Per100Plays
}#'}

\item{season_types}{vector of season types options include \itemize{
\item Regular Season
\item Pre Season
\item Playoffs
\item All Star
}}

\item{playoff_rounds}{vector of playoff rounds options include code{0:4}}

\item{is_plus_minus}{\code{TRUE} returns plus minus}

\item{n_games}{number n of last games}

\item{is_rank}{if \code{TRUE} returns rank}

\item{is_pace_adjusted}{if \code{TRUE} adjusts for pace}

\item{outcomes}{vector of outcomes options include \itemize{
\item NA
\item Wins
\item Losses
}}

\item{locations}{vector of locations options include \itemize{
\item NA
\item Home
\item Road
}}

\item{months}{vector of game months options include \code{0:12}}

\item{season_segments}{vector of season segments, options include \itemize{
\item NA
\item Post All-Star
\item Pre All-Star
}}

\item{date_from}{\code{NA} or date from}

\item{date_to}{\code{NA} or date to}

\item{opponent_ids}{vector of opponent ids}

\item{vs_confs}{vector of conferences against options include  \itemize{
\item NA
\item East
\item West
}}

\item{vs_divisions}{vector of divisions against options include \itemize{
\item NA
\item Atlantic
\item Central
\item Northwest
\item Pacific
\item Southeast
\item Southwest
}}

\item{game_segments}{vector of game segments options include \itemize{
\item NA
\item First Half
\item Second Half
\item Overtime
}}

\item{periods}{vector of periods \code{0:12}}

\item{shot_clocks}{vector of shot clock ranges options include \itemize{
\item  NA,
\item 24-22
\item 22-18 Very Early
\item 18-15 Early
\item 15-7 Average
\item 7-4 Late
\item 4-0 Very Late
\item ShotClock Off
}}

\item{last_n_games}{vector of last_n games \code{0:82}}

\item{assign_to_environment}{if \code{TRUE} assigns data to environment}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{data_frame}
}
\description{
NBA players table data
}
\examples{
get_players_tables_data(players = c("Caris LeVert", "Joe Harris"), tables =  c("year over year", "passes", "game splits"),   modes = c("PerGame", "Totals"), measures = c("Base", "Advanced"), assign_to_environment = TRUE)
}
\seealso{
Other player: \code{\link{get_bref_players_bios}},
  \code{\link{get_players_awards}},
  \code{\link{get_players_career_stats}},
  \code{\link{get_years_draft_combines}}
}
\concept{player}
