% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_game_logs.R
\name{get_seasons_rosters}
\alias{get_seasons_rosters}
\title{Get Seasons Rosters}
\usage{
get_seasons_rosters(seasons = 2000:2018, return_message = TRUE,
  nest_data = F)
}
\arguments{
\item{seasons}{vector of seasons}

\item{return_message}{if \code{TRUE} returns a message}

\item{nest_data}{if \code{TRUE} nests data}
}
\description{
Get Seasons Rosters
}
\examples{
get_seasons_rosters(200)
}
