% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons_rosters.R
\name{get_seasons_rosters}
\alias{get_seasons_rosters}
\title{NBA teams seasons rosters}
\usage{
get_seasons_rosters(seasons = NULL, return_message = TRUE,
  nest_data = F)
}
\arguments{
\item{seasons}{vector of seasons}

\item{return_message}{if \code{TRUE} returns a message}

\item{nest_data}{if \code{TRUE} nests data}
}
\value{
a `date_frame`
}
\description{
Rosters for each team of specified seasons
}
\examples{
library(nbastatR)
library(dplyr)
df_rosters <- get_seasons_rosters(2015:2018)

### Mean Age by Season and Team
df_rosters \%>\%
group_by(slugSeason, slugTeam) \%>\%
summarise(ageMean = mean(agePlayer)) \%>\%
arrange(ageMean) \%>\%
ungroup()
}
\seealso{
Other teams: \code{\link{get_drafts}},
  \code{\link{get_teams_franchise_leaders}},
  \code{\link{get_teams_players_seasons_summary_stats}},
  \code{\link{get_teams_seasons_rankings}}

Other rosters: \code{\link{get_seasons_teams_rosters}},
  \code{\link{get_team_season_roster}}
}
\concept{rosters}
\concept{teams}
