% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logs.R
\name{get_seasons_schedule}
\alias{get_seasons_schedule}
\title{Get NBA Seasons Schedule}
\usage{
get_seasons_schedule(seasons = 2018, season_types = "Regular Season",
  nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{seasons}{vector of seasons where season is year ending}

\item{return_message}{}
}
\description{
Get NBA Seasons Schedule
}
\examples{
get_seasons_schedule(seasons = c(2012, 2018))
}
