% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_game_logs.R
\name{get_seasons_teams}
\alias{get_seasons_teams}
\title{Get Seasons teams}
\usage{
get_seasons_teams(return_message = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Get Seasons teams
}
\examples{
get_seasons_teams()
}
