% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_rosters_history.R
\name{get_seasons_teams_rosters}
\alias{get_seasons_teams_rosters}
\title{Teams seasons rosters}
\usage{
get_seasons_teams_rosters(seasons = 1990:2018, nest_data = F,
  return_message = T)
}
\arguments{
\item{seasons}{vector of seasons}

\item{nest_data}{if \code{TRUE} returns nested data frame}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{data_frame}
}
\description{
Gets NBA teams rosters for specified seasons
}
\examples{
get_seasons_teams_rosters(seasons = 2010:2018, nest_data = F, return_message = T)
}
\seealso{
Other rosters: \code{\link{get_seasons_rosters}},
  \code{\link{get_team_season_roster}}
}
\concept{rosters}
