% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoops_hype_functions.R
\name{get_teams_hoopshype_salaries}
\alias{get_teams_hoopshype_salaries}
\title{HoopsHype teams players salaries}
\usage{
get_teams_hoopshype_salaries(teams = NULL, all_teams = TRUE,
  assume_player_opt_out = T, assume_team_doesnt_exercise = T,
  nest_data = F, return_message = TRUE)
}
\arguments{
\item{teams}{vector of team names}

\item{all_teams}{if \code{TRUE} returns all teams}

\item{assume_player_opt_out}{if \code{TRUE} assumes player opt out of options}

\item{assume_team_doesnt_exercise}{if \code{TRUE} assumes teams don't excerise options}

\item{nest_data}{if \code{TRUE} nests data}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a `data_frame`
}
\description{
gets detailed player salaries from HoopsHype
}
\examples{
get_teams_hoopshype_salaries(all_teams = TRUE, nest_data = F, return_message = T)
}
