% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draft_combine_data.R
\name{get_years_draft_combines}
\alias{get_years_draft_combines}
\title{get years draft combines}
\usage{
get_years_draft_combines(combine_years = 2000:2016)
}
\arguments{
\item{combine_years}{}
}
\description{
get years draft combines
}
