% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_combine.R
\name{get_years_draft_combines}
\alias{get_years_draft_combines}
\title{NBA draft combine data}
\usage{
get_years_draft_combines(years = NULL, return_message = T,
  nest_data = F)
}
\arguments{
\item{years}{vector of draft years}

\item{return_message}{if \code{TRUE} return message}

\item{nest_data}{if \code{TRUE} returns nested data_frame}
}
\value{
a \code{data_frame()}
}
\description{
Acquires NBA draft combine data
}
\examples{
get_years_draft_combines(years = c(2001:2018),
nest_data = T)
}
\seealso{
Other draft: \code{\link{get_drafts}},
  \code{\link{get_nbadraftnet_mock_drafts}}

Other player: \code{\link{get_bref_players_bios}},
  \code{\link{get_players_awards}},
  \code{\link{get_players_career_stats}},
  \code{\link{get_players_tables_data}}
}
\concept{draft}
\concept{player}
