% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_per_minute}
\alias{scale_per_minute}
\title{Summarise data per minute}
\usage{
scale_per_minute(data, scale_columns = c("pts", "fg", "ast", "tov", "blk",
  "stl", "drb", "trb", "orb", "ft", "pf", "countLayupsShooting", "countDunks",
  "hlf"))
}
\arguments{
\item{scale_columns}{}
}
\description{
Summarise data per minute
}
